@@ -7,24 +7,24 @@ description: Demonstrates how to use the fair plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  convex_bottom_bar: ^2.5.1
-  flutter_redux: ^0.6.0
+  convex_bottom_bar: ^3.0.0
+  flutter_redux: ^0.9.0
   fair:
     path: ../../fair
-  dio: ^3.0.0
-  permission_handler: 5.0.1+1
-  image_picker: ^0.6.7+22
+  dio: ^4.0.0
+  permission_handler: 6.0.1
+  image_picker: ^0.8.4+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.4.0
+  build_runner: ^2.0.0
   fair_compiler:
     path: ../../compiler
 
