@@ -24,7 +24,7 @@ class BundleGenerator extends GeneratorForAnnotation<FairPatch>
   final _elements = <String, Element>{};
 
   @override
-  Future<String> generateForAnnotatedElement(
+  Future<String?> generateForAnnotatedElement(
       Element element, ConstantReader annotation, BuildStep buildStep) async {
     if (element is! ClassElement) {
       throw InvalidGenerationSourceError(
@@ -43,7 +43,7 @@ class BundleGenerator extends GeneratorForAnnotation<FairPatch>
       if (e != element) {
         throw InvalidGenerationSourceError(
             'Each .dart file should contain only one @FairPatch Widget.\n'
-            'Both ${e.name} and ${element.name} exist inside ${e.source.uri}\n'
+            'Both ${e.name} and ${element.name} exist inside ${e.source?.uri}\n'
             'Please spilt them into separated .dart file',
             element: element);
       }
@@ -56,31 +56,31 @@ class BundleGenerator extends GeneratorForAnnotation<FairPatch>
 }
 
 class BindingGenerator extends GeneratorForAnnotation<FairBinding> {
-  StringBuffer buffer;
+  StringBuffer? buffer;
   bool generated = false;
 
   @override
-  Future<String> generateForAnnotatedElement(
+  Future<String?> generateForAnnotatedElement(
       Element element, ConstantReader annotation, BuildStep buildStep) async {
     var dir = path.join('.dart_tool', 'build', 'fairc', 'source');
     if (buffer == null) {
       buffer = StringBuffer();
-      buffer.writeln('# Generated by Fair on ${DateTime.now()}.');
+      buffer!.writeln('# Generated by Fair on ${DateTime.now()}.');
     }
 
     var resource =
-        annotation.peek('packages')?.listValue?.map((e) => e.toStringValue());
+        annotation.peek('packages')?.listValue.map((e) => e.toStringValue());
     var list = [];
     if (resource != null && resource.isNotEmpty) {
       await Future.forEach(
           resource
-              .where((p) => p.startsWith('package:') && p.endsWith('.dart')),
+              .where((p) => p?.startsWith('package:')==true && p?.endsWith('.dart')==true),
           (element) async {
-        buffer.writeln(element);
+        buffer!.writeln(element);
         var str = element
             .toString()
             .substring(0, element.toString().indexOf('\/') + 1);
-        var assetId = FairAssetId.resolve(element);
+        var assetId = FairAssetId.resolve(element as String);
         await _transitSource(buildStep, assetId, dir).then((value) {
           var elements = [];
           value.readAsLinesSync().forEach((element) {
@@ -91,17 +91,17 @@ class BindingGenerator extends GeneratorForAnnotation<FairBinding> {
                   .add(element.contains('package:') ? package : str + package);
             }
           });
-          if (elements != null && elements.isNotEmpty) {
+          if (elements.isNotEmpty) {
             list.addAll(elements);
           }
         });
       });
-      if (list != null && list.isNotEmpty) {
+      if (list.isNotEmpty) {
         await Future.forEach(
             list.where((p) => p.startsWith('package:') && p.endsWith('.dart')),
             (element) async {
-          buffer.writeln(element);
-          var assetId = FairAssetId.resolve(element);
+          buffer!.writeln(element);
+          var assetId = FairAssetId.resolve(element as String);
           await _transitSource(buildStep, assetId, dir);
         });
       }
@@ -113,7 +113,7 @@ class BindingGenerator extends GeneratorForAnnotation<FairBinding> {
     await package.writeAsString(buffer.toString());
     // FairBinding only work with class
     if (element is ClassElement) {
-      buffer.writeln('${buildStep.inputId.uri}');
+      buffer!.writeln('${buildStep.inputId.uri}');
       await _transitSource(buildStep, buildStep.inputId, dir);
     }
     if (!generated) {
@@ -143,7 +143,7 @@ class BindingGenerator extends GeneratorForAnnotation<FairBinding> {
       if (line.startsWith('import') && !line.contains(':')) {
         var asset = line.replaceFirst('import', '').replaceAll('\'', '').replaceAll(';', '').trim();
         try {
-          var lAssetId = FairAssetId.resolve(asset,from: assetId);
+          var lAssetId = FairAssetId.resolve(asset,from: assetId as FairAssetId);
           line = 'import \'package:${assetId.package}${lAssetId.path.replaceFirst('lib', '')}\';';
         } catch (e) {
           print(e);
@@ -177,8 +177,8 @@ class PackageBuilder extends Builder with FairCompiler {
             e.substring(0, end), () => e.substring(end + 1).trimRight());
       });
       return nameMap
-          .map((key, value) => MapEntry(key, packagesMap[key] + value))
-          .values;
+          .map((key, value) => MapEntry(key, "${packagesMap[key]}$value"))
+          .values.toList();
     }
     return [];
   }
@@ -188,7 +188,7 @@ class PackageBuilder extends Builder with FairCompiler {
     var assets = await buildStep
         .findAssets(Glob('lib/**.fair.ignore', recursive: true))
         .toList();
-    if (assets == null || assets.isEmpty) {
+    if (assets.isEmpty) {
       return;
     }
 
