@@ -16,7 +16,7 @@ import 'fair_check_node_map.dart';
 @AstNodeCheck('lib/fairdsl/fair_ast_gen.dart')
 class NodeCheckAstVisitor extends GeneralizingAstVisitor<Map> {
   @override
-  Map visitNode(AstNode node) {
+  Map? visitNode(AstNode node) {
     stdout.writeln('${node.runtimeType}<---->${node.toSource()}');
     return super.visitNode(node);
   }
@@ -24,31 +24,29 @@ class NodeCheckAstVisitor extends GeneralizingAstVisitor<Map> {
 
 class CustomAstVisitor extends SimpleAstVisitor<Map> {
 
-  Map _visitNode(AstNode node) {
+  Map? _visitNode(AstNode? node) {
     if (node != null) {
 
       if(!checkNode.containsKey(node.runtimeType.toString())){
-        stdout.writeln('不支持的节点${node.runtimeType}<---->${node.parent.toSource()}---->${node.toSource()} ');
+        stdout.writeln('不支持的节点${node.runtimeType}<---->${node.parent?.toSource()}---->${node.toSource()} ');
       }
       return node.accept(this);
     }
     return null;
   }
 
-  List<Map> _visitNodeList(NodeList<AstNode> nodes) {
+  List<Map> _visitNodeList(NodeList<AstNode>? nodes) {
     var maps = <Map>[];
     if (nodes != null) {
       var size = nodes.length;
       for (var i = 0; i < size; i++) {
         var node = nodes[i];
-        if (node != null) {
-          if(!checkNode.containsKey(node.runtimeType.toString())){
-            stdout.writeln('不支持的节点${node.runtimeType}<---->${node.parent.toSource()}---->${node.toSource()}');
-          }
-          var res = node.accept(this);
-          if (res != null) {
-            maps.add(res);
-          }
+        if(!checkNode.containsKey(node.runtimeType.toString())){
+          stdout.writeln('不支持的节点${node.runtimeType}<---->${node.parent?.toSource()}---->${node.toSource()}');
+        }
+        var res = node.accept(this);
+        if (res != null) {
+          maps.add(res);
         }
       }
     }
@@ -56,32 +54,32 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitMapLiteralEntry(MapLiteralEntry node) {
+  Map? visitMapLiteralEntry(MapLiteralEntry node) {
     return _buildMapLiteralEntry(_visitNode(node.key), _visitNode(node.value));
   }
 
   @override
-  Map visitSetOrMapLiteral(SetOrMapLiteral node) {
+  Map? visitSetOrMapLiteral(SetOrMapLiteral node) {
     return _buildSetOrMapLiteral(_visitNodeList(node.elements));
   }
 
   @override
-  Map visitAnnotation(Annotation node) {
+  Map? visitAnnotation(Annotation node) {
     return _buildAnnotation(_visitNode(node.name),_visitNode(node.arguments));
   }
 
   @override
-  Map visitListLiteral(ListLiteral node) {
+  Map? visitListLiteral(ListLiteral node) {
     return _buildVisitListLiteral(_visitNodeList(node.elements));
   }
 
   @override
-  Map visitInterpolationExpression(InterpolationExpression node) {
+  Map? visitInterpolationExpression(InterpolationExpression node) {
     return {'type': 'InterpolationExpression', 'id': _visitNode(node.expression)};
   }
 
   @override
-  Map visitStringInterpolation(StringInterpolation node) {
+  Map? visitStringInterpolation(StringInterpolation node) {
     return {'type':'StringInterpolation',
       'elements':_visitNodeList(node.elements),
       'sourceString':node.toSource()
@@ -89,68 +87,68 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitPostfixExpression(PostfixExpression node) {
+  Map? visitPostfixExpression(PostfixExpression node) {
     return _buildPrefixExpression(
         _visitNode(node.operand), node.operator.toString(), false);
   }
 
   @override
-  Map visitExpressionStatement(ExpressionStatement node) {
+  Map? visitExpressionStatement(ExpressionStatement node) {
     return _visitNode(node.expression);
   }
 
   @override
   //node.fields子节点类型 VariableDeclarationList
-  Map visitFieldDeclaration(FieldDeclaration node) {
+  Map? visitFieldDeclaration(FieldDeclaration node) {
     return _buildVariableDeclarationList(_visitNode(node.fields.type),
         _visitNodeList(node.fields.variables),_visitNodeList(node.metadata),node.toSource());
   }
 
   @override
-  Map visitCompilationUnit(CompilationUnit node) {
+  Map? visitCompilationUnit(CompilationUnit node) {
     return _buildAstRoot(_visitNodeList(node.declarations));
   }
 
   @override
-  Map visitBlock(Block node) {
+  Map? visitBlock(Block node) {
     return _buildBloc(_visitNodeList(node.statements));
   }
 
   @override
-  Map visitBlockFunctionBody(BlockFunctionBody node) {
+  Map? visitBlockFunctionBody(BlockFunctionBody node) {
     return _visitNode(node.block);
   }
 
   @override
-  Map visitVariableDeclaration(VariableDeclaration node) {
+  Map? visitVariableDeclaration(VariableDeclaration node) {
     return _buildVariableDeclaration(
         _visitNode(node.name), _visitNode(node.initializer));
   }
 
   @override
-  Map visitVariableDeclarationStatement(VariableDeclarationStatement node) {
+  Map? visitVariableDeclarationStatement(VariableDeclarationStatement node) {
     return _visitNode(node.variables);
   }
 
   @override
-  Map visitVariableDeclarationList(VariableDeclarationList node) {
+  Map? visitVariableDeclarationList(VariableDeclarationList node) {
     return _buildVariableDeclarationList(
         _visitNode(node.type), _visitNodeList(node.variables),_visitNodeList(node.metadata),node.toSource());
   }
 
   @override
-  Map visitSimpleIdentifier(SimpleIdentifier node) {
+  Map? visitSimpleIdentifier(SimpleIdentifier node) {
     return _buildIdentifier(node.name);
   }
 
-  @override
+  // @override
 //  Map visitBinaryExpression(BinaryExpression node) {
 //    return _buildBinaryExpression(_safelyVisitNode(node.leftOperand),
 //        _safelyVisitNode(node.rightOperand), node.operator.lexeme);
 //  }
 
   @override
-  Map visitIntegerLiteral(IntegerLiteral node) {
+  Map? visitIntegerLiteral(IntegerLiteral node) {
 
     if(node.literal.lexeme.toUpperCase().startsWith('0X')){
       return _buildStringLiteral(node.literal.lexeme);
@@ -159,57 +157,57 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitDoubleLiteral(DoubleLiteral node){
+  Map? visitDoubleLiteral(DoubleLiteral node){
     return _buildNumericLiteral(node.value);
   }
 
   @override
-  Map visitFunctionDeclaration(FunctionDeclaration node) {
+  Map? visitFunctionDeclaration(FunctionDeclaration node) {
     return _buildFunctionDeclaration(
         _visitNode(node.name), _visitNode(node.functionExpression));
   }
 
   @override
-  Map visitFunctionDeclarationStatement(FunctionDeclarationStatement node) {
+  Map? visitFunctionDeclarationStatement(FunctionDeclarationStatement node) {
     return _visitNode(node.functionDeclaration);
   }
 
   //()=>方法
   @override
-  Map visitExpressionFunctionBody(ExpressionFunctionBody node) {
+  Map? visitExpressionFunctionBody(ExpressionFunctionBody node) {
     return _buildBloc([_visitNode(node.expression)]);
   }
 
   @override
-  Map visitFunctionExpression(FunctionExpression node) {
+  Map? visitFunctionExpression(FunctionExpression node) {
     return _buildFunctionExpression(
         _visitNode(node.parameters), _visitNode(node.body),
         isAsync: node.body.isAsynchronous);
   }
 
   @override
-  Map visitSimpleFormalParameter(SimpleFormalParameter node) {
+  Map? visitSimpleFormalParameter(SimpleFormalParameter node) {
     return _buildSimpleFormalParameter(
-        _visitNode(node.type), node.identifier.name);
+        _visitNode(node.type), node.identifier?.name);
   }
 
   @override
-  Map visitFormalParameterList(FormalParameterList node) {
+  Map? visitFormalParameterList(FormalParameterList node) {
     return _buildFormalParameterList(_visitNodeList(node.parameters));
   }
 
   @override
-  Map visitTypeName(TypeName node) {
+  Map? visitTypeName(TypeName node) {
     return _buildTypeName(node.name.name);
   }
 
   @override
-  Map visitReturnStatement(ReturnStatement node) {
+  Map? visitReturnStatement(ReturnStatement node) {
     return _buildReturnStatement(_visitNode(node.expression));
   }
 
   @override
-  Map visitMethodDeclaration(MethodDeclaration node) {
+  Map? visitMethodDeclaration(MethodDeclaration node) {
     return _buildMethodDeclaration(
         _visitNode(node.name),
         _visitNode(node.parameters),
@@ -222,22 +220,22 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitNamedExpression(NamedExpression node) {
+  Map? visitNamedExpression(NamedExpression node) {
     return _buildNamedExpression(
         _visitNode(node.name), _visitNode(node.expression));
   }
 
   @override
-  Map visitPrefixedIdentifier(PrefixedIdentifier node) {
+  Map? visitPrefixedIdentifier(PrefixedIdentifier node) {
     return _buildPrefixedIdentifier(
         _visitNode(node.identifier), _visitNode(node.prefix));
   }
 
   @override
-  Map visitMethodInvocation(MethodInvocation node) {
-    Map callee;
+  Map? visitMethodInvocation(MethodInvocation node) {
+    Map? callee;
     if (node.target != null) {
-      node.target.accept(this);
+      node.target?.accept(this);
       callee = {
         'type': 'MemberExpression',
         'object': _visitNode(node.target),
@@ -251,7 +249,7 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitClassDeclaration(ClassDeclaration node) {
+  Map? visitClassDeclaration(ClassDeclaration node) {
     return _buildClassDeclaration(
         _visitNode(node.name),
         _visitNode(node.extendsClause),
@@ -262,8 +260,8 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitInstanceCreationExpression(InstanceCreationExpression node) {
-    Map callee;
+  Map? visitInstanceCreationExpression(InstanceCreationExpression node) {
+    Map? callee;
     if(node.constructorName.type.name is PrefixedIdentifier){
       var prefixedIdentifier = node.constructorName.type.name as PrefixedIdentifier;
       callee = {
@@ -279,48 +277,48 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   }
 
   @override
-  Map visitSimpleStringLiteral(SimpleStringLiteral node) {
+  Map? visitSimpleStringLiteral(SimpleStringLiteral node) {
     return _buildStringLiteral(node.value);
   }
 
   @override
-  Map visitBooleanLiteral(BooleanLiteral node) {
+  Map? visitBooleanLiteral(BooleanLiteral node) {
     return _buildBooleanLiteral(node.value);
   }
 
   @override
-  Map visitArgumentList(ArgumentList node) {
+  Map? visitArgumentList(ArgumentList node) {
     return _buildArgumentList(_visitNodeList(node.arguments));
   }
 
   @override
-  Map visitLabel(Label node) {
+  Map? visitLabel(Label node) {
     return _visitNode(node.label);
   }
 
   @override
-  Map visitExtendsClause(ExtendsClause node) {
+  Map? visitExtendsClause(ExtendsClause node) {
     return _visitNode(node.superclass);
   }
 
   @override
-  Map visitImplementsClause(ImplementsClause node) {
+  Map? visitImplementsClause(ImplementsClause node) {
     return _buildImplementsClause(_visitNodeList(node.interfaces));
   }
 
   @override
-  Map visitWithClause(WithClause node) {
+  Map? visitWithClause(WithClause node) {
     return _visitNode(node);
   }
 
   @override
-  Map visitPropertyAccess(PropertyAccess node) {
-    var expression = node.parent.toSource();
+  Map? visitPropertyAccess(PropertyAccess node) {
+    var expression = node.parent?.toSource();
     return _buildVariableExpression(expression);
   }
 
   ///根节点
-  Map _buildAstRoot(List<Map> body) {
+  Map? _buildAstRoot(List<Map> body) {
     if (body.isNotEmpty) {
       return {
         'type': 'Program',
@@ -343,14 +341,14 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
 //  };
 
   //变量声明
-  Map _buildVariableDeclaration(Map id, Map init) => {
+  Map _buildVariableDeclaration(Map? id, Map? init) => {
     'type': 'VariableDeclarator',
     'id': id,
     'init': init,
   };
 
   //变量声明列表
-  Map _buildVariableDeclarationList(Map typeAnnotation,
+  Map _buildVariableDeclarationList(Map? typeAnnotation,
       List<Map> declarations,List<Map> annotations,String source) => {
     'type': 'VariableDeclarationList',
     'typeAnnotation': typeAnnotation,
@@ -363,19 +361,19 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   Map _buildIdentifier(String name) => {'type': 'Identifier', 'name': name};
 
   //数值定义
-  Map _buildNumericLiteral(num value) => {
+  Map _buildNumericLiteral(num? value) => {
     'type': 'NumericLiteral', 'value': value
   };
 
   //函数声明
-  Map _buildFunctionDeclaration(Map id, Map expression) => {
+  Map _buildFunctionDeclaration(Map? id, Map? expression) => {
     'type': 'FunctionDeclaration',
     'id': id,
     'expression': expression,
   };
 
   //函数表达式
-  Map _buildFunctionExpression(Map params, Map body, {bool isAsync=false}) => {
+  Map _buildFunctionExpression(Map? params, Map? body, {bool isAsync=false}) => {
     'type': 'FunctionExpression',
     'parameters': params,
     'body': body,
@@ -388,7 +386,7 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   };
 
   //函数参数
-  Map _buildSimpleFormalParameter(Map type, String name) => {
+  Map _buildSimpleFormalParameter(Map? type, String? name) => {
     'type': 'SimpleFormalParameter', 'paramType': type, 'name': name
   };
 
@@ -399,14 +397,14 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
   };
 
   //返回数据定义
-  Map _buildReturnStatement(Map argument) => {
+  Map _buildReturnStatement(Map? argument) => {
     'type': 'ReturnStatement',
     'argument': argument,
   };
 
   //方法声明
   Map _buildMethodDeclaration(
-      Map id, Map parameters, Map typeParameters, Map body, Map returnType,List<Map> annotations,String source,
+      Map? id, Map? parameters, Map? typeParameters, Map? body, Map? returnType,List<Map> annotations,String source,
       {bool isAsync = false}) => {
     'type': 'MethodDeclaration',
     'id': id,
@@ -419,27 +417,27 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
     'source':source
   };
 
-  Map _buildNamedExpression(Map id, Map expression) => {
+  Map _buildNamedExpression(Map? id, Map? expression) => {
     'type': 'NamedExpression',
     'id': id,
     'expression': expression,
   };
 
-  Map _buildPrefixedIdentifier(Map identifier, Map prefix) => {
+  Map _buildPrefixedIdentifier(Map? identifier, Map? prefix) => {
     'type': 'PrefixedIdentifier',
     'identifier': identifier,
     'prefix': prefix,
   };
 
-  Map _buildMethodInvocation(Map callee, Map typeArguments, Map argumentList) => {
+  Map _buildMethodInvocation(Map? callee, Map? typeArguments, Map? argumentList) => {
     'type': 'MethodInvocation',
     'callee': callee,
     'typeArguments': typeArguments,
     'argumentList': argumentList,
   };
 
-  Map _buildClassDeclaration(Map id, Map superClause, Map implementsClause,
-      Map mixinClause, List<Map> metadata, List<Map> body) => {
+  Map _buildClassDeclaration(Map? id, Map? superClause, Map? implementsClause,
+      Map? mixinClause, List<Map> metadata, List<Map> body) => {
     'type': 'ClassDeclaration',
     'id': id,
     'superClause': superClause,
@@ -464,13 +462,13 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
     'expression': expression,
   };
 
-  Map _buildVariableExpression(String expression) =>
+  Map _buildVariableExpression(String? expression) =>
       {'type': 'VariableExpression', 'expression': expression};
 
 //  Map _buildPostfixExpression(Map operand, String operator) =>
 //      {'type': 'PostfixExpression', 'operand': operand, 'operator': operator};
 
-  Map _buildPrefixExpression(Map argument, String oprator, bool prefix) => {
+  Map _buildPrefixExpression(Map? argument, String oprator, bool prefix) => {
     'type': 'PrefixExpression',
     'argument': argument,
     'prefix': prefix,
@@ -481,7 +479,7 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
       {'type': 'ListLiteral', 'elements': literal};
 
 
-  Map _buildAnnotation(Map name,Map argumentList) =>{
+  Map _buildAnnotation(Map? name,Map? argumentList) =>{
 
     'type':'Annotation',
     'id':name,
@@ -493,7 +491,7 @@ class CustomAstVisitor extends SimpleAstVisitor<Map> {
     'elements':elements
   };
 
-  Map _buildMapLiteralEntry(Map key,Map expression)=>{
+  Map _buildMapLiteralEntry(Map? key,Map? expression)=>{
     'type':'MapLiteralEntry',
     'key':key,
     'value':expression
@@ -513,6 +511,9 @@ Future<Map> generateAstMap(String path) async {
         var compilationUnit = parseResult.unit;
         //遍历AST
         var astData = compilationUnit.accept(CustomAstVisitor());
+
+        // var encode = json.encode(astData);
+        // print(encode);
         // stdout.writeln('################gen_ast_begin######################');
         // stdout.writeln(jsonEncode(astData));
         // stdout.writeln('################gen_ast_end######################');
