@@ -53,8 +53,7 @@ Future<void> parseSdkFile(String sdkName, String filePath, String output) async 
 }
 
 Future<void> saveIntoFile(String result, String outputDir, String fileName) async {
-  var file = File('$outputDir/$fileName')
-    ..createSync(recursive: true);
+  var file = File('$outputDir/$fileName')..createSync(recursive: true);
   await file.writeAsString(result);
   await Process.start('dartfmt', ['-w', '$outputDir']);
 }
@@ -63,23 +62,20 @@ Future<String> parseDir(Directory dir) async {
   var files = await dir
       .list(recursive: true)
       .map((element) => element.absolute.path)
-      .where((event) =>
-          event.endsWith('.dart') &&
-          (transformer.skipSource.indexWhere((skip) => event.endsWith(skip)) ==
-              -1))
+      .where((event) => event.endsWith('.dart') && (transformer.skipSource.indexWhere((skip) => event.endsWith(skip)) == -1))
       .toList();
   var collection = AnalysisContextCollection(includedPaths: files);
   var components = <ComponentParts>[];
   for (var path in files) {
     var context = collection.contextFor(path);
     var w = await processFile(context, path);
-    if (w == null || w.components.isEmpty) continue;
-    components.add(w);
+    if (w?.components?.isEmpty ?? false) continue;
+    components.add(w!);
   }
   return _generateWidget(components: components);
 }
 
-Future<void> parseFlutter(String flutterSDK) async {
+Future<void> parseFlutter(String? flutterSDK) async {
   assert(flutterSDK != null);
   var frameworkAPI = [
     'packages/flutter/lib/src/widgets/',
@@ -96,11 +92,7 @@ Future<void> parseFlutter(String flutterSDK) async {
     var files = await dir
         .list(recursive: true)
         .map((element) => element.absolute.path)
-        .where((event) =>
-            event.endsWith('.dart') &&
-            (transformer.skipSource
-                    .indexWhere((skip) => event.endsWith(skip)) ==
-                -1))
+        .where((event) => event.endsWith('.dart') && (transformer.skipSource.indexWhere((skip) => event.endsWith(skip)) == -1))
         .toList();
     includedPaths.addAll(files);
   }
@@ -152,54 +144,49 @@ Future<void> _analyzeFramework(List<String> paths) async {
   for (var path in paths) {
     var context = cc.contextFor(path);
     var w = await processFile(context, path);
-    if (w == null || w.components.isEmpty) continue;
-    widgets.addAll(w.components);
+    if (w == null || w.components?.isNotEmpty!=true) continue;
+    widgets.addAll(w.components!);
     var n = _generateFileName(path);
     // name shrinking
     var index = n.indexOf('_');
     if (index != -1) {
-      var s = shrinking.entries.firstWhere(
-          (element) => n.startsWith(element.key),
-          orElse: () => MapEntry('a', 'a'));
+      var s = shrinking.entries.firstWhere((element) => n.startsWith(element.key), orElse: () => MapEntry('a', 'a'));
       n = s.value;
     }
     w.aliasGroup = n;
     components.add(w);
   }
 
-  components.sort((a, b) => a.aliasGroup.compareTo(b.aliasGroup));
+  components.sort((a, b) => a.aliasGroup?.compareTo(b.aliasGroup??'')??0);
   shrinking.forEach((key, value) {
     var c = components.where((element) => element.aliasGroup == value).toList();
     if (c.isEmpty) return;
-    var file = File('$outputPrefix/lib/src/widgets/\$\$$value.dart')
-      ..createSync(recursive: true);
+    var file = File('$outputPrefix/lib/src/widgets/\$\$$value.dart')..createSync(recursive: true);
     imports.add('\$\$$value');
-    var isCupertino = c.first.isCupertino;
-    var buffer = StringBuffer(
-        '''// This file is generated by Fair, do not edit manually!\n
+    var isCupertino = c.first.isCupertino??false;
+    var buffer = StringBuffer('''// This file is generated by Fair, do not edit manually!\n
 import 'package:flutter/${isCupertino ? "cupertino.dart" : "material.dart"}';\n''');
 
-    var allImports = c.fold(<String>{}, (p, e) {
+    var allImports = c.fold(<String>[], (List<String> p, e) {
       if (e.imports != null) {
-        p.addAll(e.imports);
+        p.addAll(e.imports??[]);
       }
       return p;
     });
     allImports.forEach((element) => buffer.writeln('import \'$element\';'));
 
-    var l = c.fold(<String>{}, (p, e) {
+    var l = c.fold(<String>[], (List<String> p, e) {
       if (e.lines != null) {
-        p.addAll(e.lines);
+        p.addAll(e.lines??[]);
       }
       return p;
     });
     var lines = l.toList();
-    lines.sort((a, b) =>
-        a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
+    lines.sort((a, b) => a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
     lines.forEach((element) => buffer.writeln('$element'));
-    var body = c.fold(<String>{}, (p, e) {
+    var body = c.fold(<String>[], (List<String> p, e) {
       if (e.body != null) {
-        p.add(e.body);
+        p.add(e.body??'');
       }
       return p;
     });
@@ -225,10 +212,9 @@ import 'package:flutter/${isCupertino ? "cupertino.dart" : "material.dart"}';\n'
   }
 
   bindingBuffer.writeln('];}');
-  var f = File('$outputPrefix/lib/src/widgets/all.dart')
-    ..createSync(recursive: true);
+  var f = File('$outputPrefix/lib/src/widgets/all.dart')..createSync(recursive: true);
   await f.writeAsString(bindingBuffer.toString());
-  var widgetCount = widgets.where((element) => element.isWidget).length;
+  var widgetCount = widgets.where((element) => element.isWidget==true).length;
   var apiCount = widgets.length - widgetCount;
   buffer.writeln('''
   const widgetCount = $widgetCount;
@@ -239,43 +225,40 @@ import 'package:flutter/${isCupertino ? "cupertino.dart" : "material.dart"}';\n'
     buffer.writeln('\'${element.name}\': ${element.isWidget},');
   });
   buffer.writeln('};');
-  await File('$outputPrefix/lib/src/widgets/version.dart')
-      .writeAsString(buffer.toString());
+  await File('$outputPrefix/lib/src/widgets/version.dart').writeAsString(buffer.toString());
 }
 
-Future<String> _generateWidget(
-    {List<ComponentParts> components,
-    Future<ComponentParts> Function() builder}) async {
-  var c = components ?? [await builder()];
+Future<String> _generateWidget({List<ComponentParts?>? components, Future<ComponentParts> Function()? builder}) async {
+  var c = components ?? (builder==null?null:[await builder()]);
   if (c == null || c.first == null) return '';
-  var isCupertino = c.first.isCupertino;
+  var isCupertino = c.first!.isCupertino??false;
   var buffer = StringBuffer('''
   import 'package:flutter/${isCupertino ? "cupertino.dart" : "material.dart"}';
   import 'package:fair/fair.dart';
   import 'package:fair_version/fair_version.dart';
   ''');
 
-  var allImports = c.fold(<String>{}, (p, e) {
-    if (e.imports != null) {
-      p.addAll(e.imports);
+
+  var allImports = c.fold(<String>[], (List<String> p,e) {
+    if (e?.imports != null) {
+      p.addAll(e?.imports??[]);
     }
     return p;
   });
   allImports.forEach((element) => buffer.write('import \'$element\';\n'));
 
-  var l = c.fold(<String>{}, (p, e) {
-    if (e.lines != null) {
-      p.addAll(e.lines);
+  var l = c.fold(<String>[], (List<String> p, e) {
+    if (e?.lines != null) {
+      p.addAll(e?.lines??[]);
     }
     return p;
   });
   var lines = l.toList();
-  lines.sort(
-      (a, b) => a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
+  lines.sort((a, b) => a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
   lines.forEach((element) => buffer.write('$element\n'));
-  var body = c.fold(<String>{}, (p, e) {
-    if (e.body != null) {
-      p.add(e.body);
+  var body = c.fold(<String>[], (List<String> p, e) {
+    if (e?.body != null) {
+      p.add(e?.body??'');
     }
     return p;
   });
@@ -291,13 +274,14 @@ Future<String> _generateWidget(
     }
   ''');
 
-
   buffer.write('''
   @override
   Map<String, bool> mapping() {
       return const {
   ''');
-  c.forEach((element) => element.components.forEach((e) {buffer.write('\'${e.name}\': ${e.isWidget},\n'); }));
+  c.forEach((element) => element?.components?.forEach((e) {
+        buffer.write('\'${e.name}\': ${e.isWidget},\n');
+      }));
   buffer.write('''
       };
       }
@@ -306,28 +290,27 @@ Future<String> _generateWidget(
   return buffer.toString();
 }
 
-Future<void> _generateSdkFile(String sdkName, String filePath, String output) async {
+Future _generateSdkFile(String? sdkName, String filePath, String output) async {
   var collection = AnalysisContextCollection(includedPaths: [filePath]);
   var context = collection.contextFor(filePath);
-  var c =  await processFile(context, filePath, analysisExports: true);
+  var c = await processFile(context, filePath, analysisExports: true);
   if (c == null) return '';
-  var isCupertino = c.isCupertino;
+  var isCupertino = c.isCupertino??false;
   var buffer = StringBuffer('''
   import 'package:flutter/${isCupertino ? "cupertino.dart" : "material.dart"}';
   ''');
   var fileName = basename(filePath);
   var clzName = fileName.replaceAll('.dart', '');
-  if(sdkName?.isNotEmpty != true){
+  if (sdkName?.isNotEmpty != true) {
     sdkName = clzName;
   }
   buffer.write('''
   import 'package:$sdkName/$fileName';
   ''');
 
-  var lines = c.lines.toList();
-  lines.sort(
-          (a, b) => a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
-  lines.forEach((element) => buffer.write('$element\n'));
+  var lines = c.lines?.toList();
+  lines?.sort((a, b) => a.startsWith('import') ? -1 : (b.startsWith('import') ? 1 : 0));
+  lines?.forEach((element) => buffer.write('$element\n'));
   var body = c.body;
   buffer.write('''
   var ${clzName}_component = {
@@ -338,21 +321,18 @@ Future<void> _generateSdkFile(String sdkName, String filePath, String output) as
   await saveIntoFile(buffer.toString(), output, 'fair_${clzName}.dart');
 }
 
-Future<ComponentParts> processFile(AnalysisContext context, String path,
-    {bool analysisExports = false}) async {
+Future<ComponentParts?> processFile(AnalysisContext context, String path, {bool analysisExports = false}) async {
   var session = context.currentSession;
-  UnitElementResult result = await session.getUnitElement(path);
+  var result = await session.getUnitElement(path) as UnitElementResult;
   var element = result.element;
-  var elementsList = <CompilationUnitElement>[element];
+  var elementsList = <CompilationUnitElement?>[element];
   //analysis exports file
-  if(analysisExports){
+  if (analysisExports) {
     //Most of SDK entrance files export other files.
     //We need analysis all of them when we are compiling.
     var exports = result.element.enclosingElement.exports;
-    if(exports != null){
-      var exportsUnits = exports?.map((e) => e.exportedLibrary.definingCompilationUnit)?.toList();
-      elementsList.addAll(exportsUnits);
-    }
+    var exportsUnits = exports.map((e) => e.exportedLibrary?.definingCompilationUnit).toList();
+    elementsList.addAll(exportsUnits);
   }
 
   var exposedAPI = <ClassExposed>[];
@@ -360,7 +340,7 @@ Future<ComponentParts> processFile(AnalysisContext context, String path,
     exposedAPI.addAll(_visit(element, analysisExports));
   });
 
-  var count = exposedAPI?.length ?? 0;
+  var count = exposedAPI.length;
   print('😀 $count widgets found inside $path');
   var isCupertino = path.contains('/cupertino/');
 
@@ -370,30 +350,27 @@ Future<ComponentParts> processFile(AnalysisContext context, String path,
   var lines = transformer.getLines(clzName, isCupertino);
   // constructor api
   var hasConstructor = false;
-  exposedAPI?.forEach((e) {
+  exposedAPI.forEach((e) {
     var hasNamedConstructor = false;
     var defaultCache = <String, dynamic>{};
-    e.constructor.forEach((element) {
-      if (element.parameters != null && element.parameters.isNotEmpty) {
+    e.constructor?.forEach((element) {
+      if (element.parameters != null && (element.parameters?.isNotEmpty ?? false)) {
         hasNamedConstructor = true;
       }
-      defaultCache =
-          _writeMethod(buffer, element.name, element, defaultCache);
+      defaultCache = _writeMethod(buffer, element.name, element, defaultCache);
     });
-    e.functionParameters.forEach((element) {
-      _writeFunctionParameter(
-          buffer, element.toString(), element, defaultCache);
+    e.functionParameters?.forEach((element) {
+      _writeFunctionParameter(buffer, element.toString(), element, defaultCache);
     });
     if (hasNamedConstructor) {
-      e.fields.forEach((element) {
+      e.fields?.forEach((element) {
         var field = element;
-        buffer.write(
-            '\'${field.name}.${field.field}\': ${field.name}.${field.field},');
+        buffer.write('\'${field.name}.${field.field}\': ${field.name}.${field.field},');
       });
-    } else if (e.fields.isNotEmpty) {
+    } else if (e.fields?.isNotEmpty == true) {
       var name = e.name;
       buffer.write('\'$name\': {');
-      e.fields.forEach((e) {
+      e.fields?.forEach((e) {
         var field = e;
         buffer.write('\'${field.field}\': ${field.name}.${field.field},');
       });
@@ -401,22 +378,20 @@ Future<ComponentParts> processFile(AnalysisContext context, String path,
     }
 
     e.staticMethods?.forEach((element) {
-      defaultCache =
-          _writeMethod(buffer, element.name, element, defaultCache);
+      defaultCache = _writeMethod(buffer, element.name, element, defaultCache);
     });
   });
-  if (!hasConstructor && exposedAPI != null && exposedAPI.isEmpty) return null;
+  if (!hasConstructor && exposedAPI.isEmpty) return null;
   return ComponentParts(
     exposedAPI
-        ?.map((e) => Component(
+        .map((e) => Component(
               e.name,
-              e.constructor != null && e.constructor.isNotEmpty
-                  ? e.constructor[0].isWidget
-                  : ((e.fields?.isNotEmpty == true && e.fields.elementAt(0)?.isWidget==true)
-                  || (e.staticMethods?.isNotEmpty == true && e.staticMethods.elementAt(0)?.isWidget==true)
-              ),
+              e.constructor != null && e.constructor?.isNotEmpty == true
+                  ? e.constructor![0].isWidget
+                  : ((e.fields?.isNotEmpty == true && e.fields!.elementAt(0).isWidget == true) ||
+                      (e.staticMethods?.isNotEmpty == true && e.staticMethods!.elementAt(0).isWidget == true)),
             ))
-        ?.toList(growable: false),
+        .toList(growable: false),
     buffer.toString(),
     imports: imports,
     lines: lines,
@@ -426,13 +401,13 @@ Future<ComponentParts> processFile(AnalysisContext context, String path,
 }
 
 class ComponentParts {
-  final bool isCupertino;
-  final List<String> imports;
-  final List<String> lines;
-  final List<Component> components;
-  String aliasGroup;
-  final String body;
-  final String clzName;
+  final bool? isCupertino;
+  final List<String>? imports;
+  final List<String>? lines;
+  final List<Component>? components;
+  String? aliasGroup;
+  final String? body;
+  final String? clzName;
 
   ComponentParts(
     this.components,
@@ -445,24 +420,22 @@ class ComponentParts {
 }
 
 class Component {
-  final String name;
-  final bool isWidget;
+  final String? name;
+  final bool? isWidget;
 
   Component(this.name, this.isWidget);
 }
 
 String _generateFileName(String path) {
-  var fileName =
-      path.split('packages/flutter/lib/src/')[1].replaceAll('/', '_');
+  var fileName = path.split('packages/flutter/lib/src/')[1].replaceAll('/', '_');
   return fileName;
 }
 
-Map<String, dynamic> _writeMethod(StringBuffer buffer, String name,
-    Method element, Map<String, dynamic> defaultCache) {
+Map<String, dynamic> _writeMethod(StringBuffer buffer, String? name, Method element, Map<String, dynamic> defaultCache) {
   buffer.write('\'$name\': (props) => $name(');
-  if (element.parameters != null && element.parameters.isNotEmpty) {
-    for (var i = 0; i < element.parameters.length; i++) {
-      var p = element.parameters[i];
+  if (element.parameters != null && (element.parameters?.isNotEmpty ?? false)) {
+    for (var i = 0; i < element.parameters!.length; i++) {
+      var p = element.parameters![i];
       if (!transformer.isPropSupported(p.name)) continue;
       String prop;
       // JSON属性中可能存在，浮点写成整型，需要兼容
@@ -472,43 +445,40 @@ Map<String, dynamic> _writeMethod(StringBuffer buffer, String name,
       // 处理List类型数据问题
       var cName = '';
       if (isList) {
-        if (p.displayName != null && p.displayName.contains('<')) {
-          cName = '<'+p.displayName.split('<')[1].split('>')[0]+'>';
+        if (p.displayName != null && p.displayName!.contains('<')) {
+          cName = '<' + p.displayName!.split('<')[1].split('>')[0] + '>';
         }
       }
 
       if (p.defaultValueCode != null) {
-        if(isDouble){
+        if (isDouble) {
           prop = 'props[\'${p.name}\']?.toDouble() ?? ${p.defaultValueCode},';
-        }else if(isList){
+        } else if (isList) {
           prop = 'as${cName}(props[\'${p.name}\']) ?? ${p.defaultValueCode},';
-        }else{
+        } else {
           prop = 'props[\'${p.name}\'] ?? ${p.defaultValueCode},';
         }
 
-        defaultCache[p.name] = p.defaultValueCode;
+        defaultCache[p.name ?? ''] = p.defaultValueCode;
       } else if (defaultCache[p.name] != null) {
         prop = isDouble
             ? 'props[\'${p.name}\']?.toDouble() ?? ${defaultCache[p.name]},'
-            : isList? 'as${cName}(props[\'${p.name}\']) ?? ${defaultCache[p.name]},'
-            : 'props[\'${p.name}\'] ?? ${defaultCache[p.name]},';
-        print(
-            '💕 using cached default value  ${element.name} ${p.name}=> ${defaultCache[p.name]}');
+            : isList
+                ? 'as${cName}(props[\'${p.name}\']) ?? ${defaultCache[p.name]},'
+                : 'props[\'${p.name}\'] ?? ${defaultCache[p.name]},';
+        print('💕 using cached default value  ${element.name} ${p.name}=> ${defaultCache[p.name]}');
       } else {
         prop = isDouble
             ? 'props[\'${p.name}\']?.toDouble(),'
-            : isList? 'as${cName}(props[\'${p.name}\']),'
-            : 'props[\'${p.name}\'],';
+            : isList
+                ? 'as${cName}(props[\'${p.name}\']),'
+                : 'props[\'${p.name}\'],';
       }
       var namedDeclare = '${p.name}: $prop';
-      var positionDeclare =
-          isDouble ? 'props[\'pa\'][$i]?.toDouble(),' : 'props[\'pa\'][$i],';
-      buffer.write(p.isNamed ? namedDeclare : positionDeclare);
+      var positionDeclare = isDouble ? 'props[\'pa\'][$i]?.toDouble(),' : 'props[\'pa\'][$i],';
+      buffer.write(p.isNamed == true ? namedDeclare : positionDeclare);
     }
-    var params = element.parameters.fold(
-        '',
-        (value, p) =>
-            value + (p.isNamed ? '${p.type} ${p.name}, ' : '${p.name}, '));
+    var params = element.parameters?.fold('', (String value, p) => ((value ?? '') + (p.isNamed == true ? '${p.type} ${p.name}, ' : '${p.name}, ')));
     print('➡️ $name({$params})');
   } else {
     print('➡️ $name()');
@@ -517,8 +487,7 @@ Map<String, dynamic> _writeMethod(StringBuffer buffer, String name,
   return defaultCache;
 }
 
-Map<String, dynamic> _writeFunctionParameter(StringBuffer buffer, String name,
-    FunctionParameter element, Map<String, dynamic> defaultCache) {
+Map<String, dynamic> _writeFunctionParameter(StringBuffer buffer, String name, FunctionParameter element, Map<String, dynamic> defaultCache) {
   buffer.write('\'$name\': ');
   if (element.typeArgument != null) {
     buffer.write('<${element.typeArgument}>');
@@ -526,16 +495,13 @@ Map<String, dynamic> _writeFunctionParameter(StringBuffer buffer, String name,
   buffer.write('(props) => ');
   buffer.write('(');
 
-  if (element.parameters != null && element.parameters.isNotEmpty) {
-    for (var i = 0; i < element.parameters.length; i++) {
-      var p = element.parameters[i];
+  if (element.parameters != null && element.parameters!.isNotEmpty) {
+    for (var i = 0; i < element.parameters!.length; i++) {
+      var p = element.parameters![i];
       buffer.write('${p.type} ${p.name}, ');
     }
 
-    var params = element.parameters.fold(
-        '',
-        (value, p) =>
-            value + (p.isNamed ? '${p.type} ${p.name}, ' : '${p.name}, '));
+    var params = element.parameters?.fold('', (String value, p) => value + (p.isNamed==true ? '${p.type} ${p.name}, ' : '${p.name}, '));
     print('➡️ $name({$params})');
   } else {
     print('➡️ $name()');
@@ -553,16 +519,13 @@ Map<String, dynamic> _writeFunctionParameter(StringBuffer buffer, String name,
 
 var transformer = TransformProxy();
 
-bool _matchType(InterfaceType type, List<String> widgets,
-    {ClassElement classElement}) {
+bool _matchType(InterfaceType? type, List<String>? widgets, {ClassElement? classElement}) {
   if (type == null) return false;
-  var hit = widgets.indexWhere((element) => element == type.name) != -1;
-  return hit ||
-      _tryInternalCheck(type, widgets, classElement) ||
-      _matchType(type.superclass, widgets);
+  var hit = (widgets ?? []).indexWhere((element) => element == type.name) != -1;
+  return hit || _tryInternalCheck(type, widgets, classElement) || _matchType(type.superclass, widgets);
 }
 
-bool _tryInternalCheck(InterfaceType type, List<String> widgets, classElement) {
+bool _tryInternalCheck(InterfaceType type, List<String>? widgets, classElement) {
   // try {
   //   if (type.superclass != null &&
   //       type.superclass.name == 'Object' &&
@@ -582,10 +545,10 @@ bool _tryInternalCheck(InterfaceType type, List<String> widgets, classElement) {
 }
 
 class ClassExposed extends Exposed {
-  final List<Constructor> constructor;
-  final List<ConstField> fields;
-  final List<Method> staticMethods;
-  final List<FunctionParameter> functionParameters;
+  final List<Constructor>? constructor;
+  final List<ConstField>? fields;
+  final List<Method>? staticMethods;
+  final List<FunctionParameter>? functionParameters;
 
   ClassExposed(
     String name, {
@@ -595,76 +558,71 @@ class ClassExposed extends Exposed {
     this.functionParameters,
   }) : super(name);
 
-  int get size =>
-      constructor.length + fields.length + functionParameters.length;
+  int get size => (constructor?.length ?? 0) + (fields?.length ?? 0) + (functionParameters?.length ?? 0);
 }
 
 abstract class Exposed {
-  final String name;
+  final String? name;
 
   Exposed(this.name);
 }
 
 class Constructor extends Method {
-
-  Constructor(String name, {List<Parameter> parameters, bool isWidget})
-      : super(name, parameters: parameters, isWidget: isWidget);
+  Constructor(String name, {List<Parameter>? parameters, bool? isWidget}) : super(name, parameters: parameters, isWidget: isWidget);
 
   @override
   String toString() {
-    return name;
+    return name ?? '';
   }
 }
 
 class Method extends Exposed {
-  final List<Parameter> parameters;
-  final bool isWidget;
+  final List<Parameter>? parameters;
+  final bool? isWidget;
 
-  Method(String name, {this.parameters, this.isWidget = true})
-      : super(name);
+  Method(String name, {this.parameters, this.isWidget = true}) : super(name);
 
   @override
   String toString() {
-    return name;
+    return name ?? '';
   }
 }
 
 class ConstField extends Exposed {
-  final String field;
-  final bool isWidget;
+  final String? field;
+  final bool? isWidget;
 
   ConstField(String name, {this.field, this.isWidget = true}) : super(name);
 
   @override
   String toString() {
-    return name + '.' + field;
+    return '$name.$field';
   }
 }
 
 class FunctionParameter extends Exposed {
-  final String className;
-  final List<Parameter> parameters;
-  final String returnType;
-  final String typeArgument;
+  final String? className;
+  final List<Parameter>? parameters;
+  final String? returnType;
+  final String? typeArgument;
 
-  FunctionParameter(String name,
-      {this.className, this.parameters, this.returnType, this.typeArgument})
-      : super(name);
+  FunctionParameter(String name, {this.className, this.parameters, this.returnType, this.typeArgument}) : super(name);
 
   @override
   String toString() {
-    return className + '#' + name;
+    return '$className#$name';
   }
 }
 
 class Parameter {
-  final String name;
-  final String type;
-  final bool isNamed;
-  final bool isOptional;
+  final String? name;
+  final String? type;
+  final bool? isNamed;
+  final bool? isOptional;
+
   //暂时为了转化List
-  final String displayName;
-  final String defaultValueCode;
+  final String? displayName;
+  final String? defaultValueCode;
 
   Parameter({
     this.name,
@@ -677,52 +635,40 @@ class Parameter {
 }
 
 bool _invalidElement(Element element) {
-  if (!element.isPublic ||
-      element.hasDeprecated ||
-      element.hasVisibleForTesting) {
+  if (!element.isPublic || element.hasDeprecated || element.hasVisibleForTesting) {
     return true;
   }
   // hasVisibleForTesting 与 isVisibleForTesting都不能正确识别注解visibleForTesting
   var meta = element.metadata;
-  var isVisibleForTesting = meta.isNotEmpty
-      ? meta.where((element) =>
-          element.isVisibleForTesting ||
-          element.toSource() == '@visibleForTesting')
-      : null;
+  var isVisibleForTesting =
+      meta.isNotEmpty ? meta.where((element) => element.isVisibleForTesting || element.toSource() == '@visibleForTesting') : null;
   return isVisibleForTesting != null && isVisibleForTesting.isNotEmpty;
 }
 
 ///
 /// We need to compile all of constructions when compile the SDK files.
 ///
-List<ClassExposed> _visit(CompilationUnitElement unitElement,[bool isSdk = false]) {
+List<ClassExposed> _visit(CompilationUnitElement? unitElement, [bool isSdk = false]) {
+  if (unitElement == null) return <ClassExposed>[];
   var exposed = <ClassExposed>[];
   // 枚举与class不同
   var apis = [...unitElement.types, ...unitElement.enums];
   for (var classElement in apis) {
-    if (classElement.isAbstract ||
-        _invalidElement(classElement) ||
-        classElement.isSynthetic) {
+    if (classElement.isAbstract || _invalidElement(classElement) || classElement.isSynthetic) {
       print('skip ' + classElement.name);
       continue;
     }
     var constructors = <Constructor>[];
     var functionParameters = <FunctionParameter>[];
-    var isWidget = classElement.thisType is InterfaceType
-        ? _matchType(classElement.thisType, baseWidget,
-            classElement: classElement)
-        : false;
-    var isAPI = classElement.thisType is InterfaceType
-        ? _matchType(classElement.thisType, baseAPI, classElement: classElement)
-        : false;
+    var isWidget = classElement.thisType is InterfaceType ? _matchType(classElement.thisType, baseWidget, classElement: classElement) : false;
+    var isAPI = classElement.thisType is InterfaceType ? _matchType(classElement.thisType, baseAPI, classElement: classElement) : false;
     if (isWidget || isAPI || isSdk) {
       print('${classElement.name} 😀');
       print('constructors ➡️');
       for (var constructorElement in classElement.constructors) {
         if (!_invalidElement(constructorElement)) {
           String name;
-          if (constructorElement.name == null ||
-              constructorElement.name == '') {
+          if (constructorElement.name == '') {
             name = classElement.name;
           } else {
             name = '${classElement.name}.${constructorElement.name}';
@@ -730,8 +676,7 @@ List<ClassExposed> _visit(CompilationUnitElement unitElement,[bool isSdk = false
           print(' $name');
           // params
           var parameters = <Parameter>[];
-          if ((constructorElement?.parameters?.isNotEmpty) ?? false) {
-
+          if (constructorElement.parameters.isNotEmpty) {
             constructorElement.parameters.forEach((e) {
               parameters.add(Parameter(
                 type: e.type.name,
@@ -769,8 +714,7 @@ List<ClassExposed> _visit(CompilationUnitElement unitElement,[bool isSdk = false
             });
           }
 
-          constructors.add(
-              Constructor(name, parameters: parameters, isWidget: isWidget));
+          constructors.add(Constructor(name, parameters: parameters, isWidget: isWidget));
         }
       }
       print('constructors ⬅️️');
@@ -780,33 +724,27 @@ List<ClassExposed> _visit(CompilationUnitElement unitElement,[bool isSdk = false
       if (!_invalidElement(fieldElement) && fieldElement.isConst) {
         print('  ❤️${classElement.name}.${fieldElement.name}');
         // 复用class类型（可能不一致）
-        fields.add(ConstField(classElement.name,
-            field: fieldElement.name, isWidget: isWidget));
+        fields.add(ConstField(classElement.name, field: fieldElement.name, isWidget: isWidget));
       }
     }
 
     var staticMethods = <Method>[];
-    for (var methodElement in classElement.methods){
-      String name ;
+    for (var methodElement in classElement.methods) {
+      String name;
       if (!_invalidElement(methodElement) && methodElement.isStatic) {
         name = '${classElement.name}.${methodElement.name}';
         print(' 👉$name');
       } else {
         continue;
       }
-      List<Parameter> parameters;
-      if (methodElement.parameters.isNotEmpty){
+      List<Parameter>? parameters;
+      if (methodElement.parameters.isNotEmpty) {
         parameters = methodElement.parameters
-            .map((e) => Parameter(
-          type: e.type.name,
-          name: e.name,
-          isNamed: e.isNamed,
-          isOptional: e.isOptional,
-          defaultValueCode: e.defaultValueCode
-        )).toList(growable: false);
+            .map(
+                (e) => Parameter(type: e.type.name, name: e.name, isNamed: e.isNamed, isOptional: e.isOptional, defaultValueCode: e.defaultValueCode))
+            .toList(growable: false);
       }
-      staticMethods.add(
-        Method(name, parameters: parameters, isWidget: isWidget));
+      staticMethods.add(Method(name, parameters: parameters, isWidget: isWidget));
     }
 
     if (constructors.isNotEmpty || fields.isNotEmpty || staticMethods.isNotEmpty) {
