@@ -6,81 +6,79 @@
 
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
-
 import '../module/module_registry.dart';
 import '../render/expression.dart';
 import '../type.dart';
 
 class BindingData {
   // none-null
-  final FairModuleRegistry modules;
-  final Map<String, Function> _functions;
-  final Map<String, PropertyValue> _values;
-  Map data;
+  final FairModuleRegistry? modules;
+  final Map<String, Function>? _functions;
+  final Map<String, PropertyValue>? _values;
+  Map? data;
   final Set<LifeCircle> _boundValue = {};
 
   BindingData(this.modules,
       {this.data,
-      Map<String, Function> functions,
-      Map<String, PropertyValue> values, Map<String, PropertyValue> delegateValues})
+      Map<String, Function>? functions,
+      Map<String, PropertyValue>? values, Map<String, PropertyValue>? delegateValues})
       : _functions = functions,
         _values = values,
         super();
 
   dynamic bindDataOf(String key) {
-    if (_values != null && _values[key] != null) {
-      return Function.apply(_values[key], null);
-    } else if (data != null && data[key] != null) {
-      return data[key];
+    if (_values != null && _values![key] != null) {
+      return Function.apply(_values![key]!, null);
+    } else if (data != null && data![key] != null) {
+      return data![key];
     }
     return functionOf(key);
   }
 
-  dynamic runFunctionOf(String funcName, {String exp}) {
-    if (_functions[funcName] == null) {
-      var result = _functions['runtimeInvokeMethodSync'](funcName);
+  dynamic runFunctionOf(String funcName, {String? exp}) {
+    if (_functions?[funcName] == null) {
+      var result = _functions?['runtimeInvokeMethodSync']?.call(funcName);
       var value = jsonDecode(result);
       return value['result']['result'];
     } else {
-      return _functions[funcName];
+      return _functions?[funcName];
     }
   }
 
   dynamic bindFunctionOf(String funcName) {
-    if (_functions[funcName] == null) {
-      return ([props]) => _functions['runtimeInvokeMethod'](funcName, props);
+    if (_functions?[funcName] == null) {
+      return ([props]) => _functions?['runtimeInvokeMethod']?.call(funcName, props);
     } else {
-      return _functions[funcName];
+      return _functions?[funcName];
     }
   }
 
   dynamic bindRuntimeValueOf(String name) {
     // _delegateValues优先级高于JS，如果要使用JS的变量，需要重命名变量
-    if (_values[name] == null) {
-        var result = _functions['runtimeParseVar']({name: ''});
+    if (_values?[name] == null) {
+        var result = _functions?['runtimeParseVar']?.call({name: ''});
         var value = jsonDecode(result);
         if (value['result'][name] != null) {
           return value['result'][name];
         }
     } else {
-      if (_values != null && _values[name] != null) {
-        return Function.apply(_values[name], null);
+      if (_values != null && _values?[name] != null) {
+        return Function.apply(_values![name]!, null);
       }
     }
     return null;
   }
 
-  Function functionOf(String key) {
-    return _functions != null ? _functions[key] : null;
+  Function? functionOf(String key) {
+    return _functions != null ? _functions![key] : null;
   }
 
-  PropertyValue valueOf(String key) {
-    return _values != null ? _values[key] : null;
+  PropertyValue? valueOf(String key) {
+    return _values != null ? _values![key] : null;
   }
 
   dynamic dataOf(String key) {
-    return data != null ? data[key] : null;
+    return data != null ? data![key] : null;
   }
 
   void addBindValue(LifeCircle value) {
