@@ -15,11 +15,11 @@ import 'fair_decoder.dart';
 import 'global_state.dart';
 
 class FairBundle {
-  BundleLoader _provider;
+  BundleLoader? _provider;
 
   BundleLoader obtain(BuildContext context) {
-    _provider ??= FairApp.of(context).bundleProvider ?? _DefaultProvider();
-    return _provider;
+    _provider ??= FairApp.of(context)?.bundleProvider ?? _DefaultProvider();
+    return _provider!;
   }
 }
 
@@ -29,37 +29,37 @@ class _DefaultProvider extends BundleLoader {
   static const FLEX = '.bin';
 
   @override
-  Future<Map> onLoad(String path, FairDecoder decoder,
-      {bool cache = true, Map<String, String> h}) {
+  Future<Map?> onLoad(String? path, FairDecoder decoder,
+      {bool cache = true, Map<String, String>? h}) {
     bool isFlexBuffer;
-    if (path.endsWith(FLEX)) {
+    if (path?.endsWith(FLEX)==true) {
       isFlexBuffer = true;
-    } else if (path.endsWith(JSON)) {
+    } else if (path?.endsWith(JSON)==true) {
       isFlexBuffer = false;
     } else {
       throw ArgumentError(
           'unknown format, please use either $JSON or $FLEX;\n $path');
     }
-    if (path.startsWith('http')) {
+    if (path?.startsWith('http')==true) {
       return _http(path, isFlexBuffer, headers: h, decode: decoder);
     }
     return _asset(path, isFlexBuffer, cache: cache, decode: decoder);
   }
 
-  Future<Map> _asset(String key, bool isFlexBuffer,
-      {bool cache = true, FairDecoder decode}) async {
+  Future<Map?> _asset(String? key, bool isFlexBuffer,
+      {bool cache = true, FairDecoder? decode}) async {
     var watch = Stopwatch()..start();
-    int end, end2;
-    Map map;
+    int? end, end2;
+    Map? map;
     if (isFlexBuffer) {
-      var data = await rootBundle.load(key);
+      var data = await rootBundle.load(key??'');
       end = watch.elapsedMilliseconds;
-      map = await decode.decode(data.buffer.asUint8List(), isFlexBuffer);
+      map = await decode?.decode(data.buffer.asUint8List(), isFlexBuffer);
       end2 = watch.elapsedMilliseconds;
     } else {
-      var data = await rootBundle.loadString(key, cache: cache);
+      var data = await rootBundle.loadString(key??'', cache: cache);
       end = watch.elapsedMilliseconds;
-      map = await decode.decode(data, isFlexBuffer);
+      map = await decode?.decode(data, isFlexBuffer);
       end2 = watch.elapsedMilliseconds;
     }
 
@@ -67,20 +67,19 @@ class _DefaultProvider extends BundleLoader {
     return map;
   }
 
-  Future<Map> _http(String url, bool isFlexBuffer,
-      {Map<String, String> headers, FairDecoder decode}) async {
+  Future<Map?> _http(String? url, bool isFlexBuffer,
+      {Map<String, String>? headers, FairDecoder? decode}) async {
     var start = DateTime.now().millisecondsSinceEpoch;
-    var response = await client.get(url, headers: headers);
+    var response = await client.get(Uri.parse(url??''), headers: headers);
     var end = DateTime.now().millisecondsSinceEpoch;
     if (response.statusCode != 200) {
       throw FlutterError('code=${response.statusCode}, unable to load : $url');
     }
     var data = response.bodyBytes;
-    if (data == null) {
-      throw FlutterError('bodyBytes=null, unable to load : $url');
-    }
-    Map map;
-    map = await decode.decode(data, isFlexBuffer);
+    // if (data == null) {
+    //   throw FlutterError('bodyBytes=null, unable to load : $url');
+    // }
+    var map = await decode?.decode(data, isFlexBuffer);
     var end2 = DateTime.now().millisecondsSinceEpoch;
     log('[Fair] load $url, time: ${end - start} ms, json parsing time: ${end2 - end} ms');
     return map;
