@@ -8,11 +8,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 class WarningWidget extends StatelessWidget {
-  final String name;
-  final String url;
+  final String? name;
+  final String? url;
   final dynamic error;
 
-  const WarningWidget({Key key, this.name, this.url, this.error})
+  const WarningWidget({Key? key, this.name, this.url, this.error})
       : super(key: key);
 
   @override
