@@ -12,16 +12,16 @@ import 'types.dart';
 
 /// The main builder class for creation of a FlexBuffer.
 class Builder {
-  ByteData _buffer;
-  List<_StackValue> _stack;
-  List<_StackPointer> _stackPointers;
-  int _offset;
-  bool _finished;
-  Map<String, _StackValue> _stringCache;
-  Map<String, _StackValue> _keyCache;
-  Map<_KeysHash, _StackValue> _keyVectorCache;
-  Map<int, _StackValue> _indirectIntCache;
-  Map<double, _StackValue> _indirectDoubleCache;
+  late ByteData _buffer;
+  late List<_StackValue> _stack;
+  late List<_StackPointer> _stackPointers;
+  late int _offset;
+  late bool _finished;
+  late Map<String, _StackValue> _stringCache;
+  late Map<String, _StackValue> _keyCache;
+  late Map<_KeysHash, _StackValue> _keyVectorCache;
+  late Map<int, _StackValue> _indirectIntCache;
+  late Map<double, _StackValue> _indirectDoubleCache;
 
   /// Instantiate the builder if you intent to gradually build up the buffer by calling
   /// add... methods and calling [finish] to receive the the resulting byte array.
@@ -52,7 +52,7 @@ class Builder {
     return byteData.buffer;
   }
 
-  void _add(Object value) {
+  void _add(Object? value) {
     if (value == null) {
       addNull();
     } else if (value is bool) {
@@ -88,7 +88,7 @@ class Builder {
   /// Specifically useful when building up a vector where values can be null.
   void addNull() {
     _integrityCheckOnValueAddition();
-    _stack.add(_StackValue.withNull());
+    _stack?.add(_StackValue.withNull());
   }
 
   /// Adds a string value.
@@ -112,8 +112,8 @@ class Builder {
   /// Adds a string value.
   void addString(String value) {
     _integrityCheckOnValueAddition();
-    if (_stringCache.containsKey(value)) {
-      _stack.add(_stringCache[value]);
+    if (_stringCache.containsKey(value) == true) {
+      _stack.add(_stringCache[value]!);
       return;
     }
     final utf8String = utf8.encode(value);
@@ -125,10 +125,9 @@ class Builder {
     final newOffset = _newOffset(length + 1);
     _pushBuffer(utf8String);
     _offset = newOffset;
-    final stackValue =
-        _StackValue.withOffset(stringOffset, ValueType.String, bitWidth);
-    _stack.add(stackValue);
-    _stringCache[value] = stackValue;
+    final stackValue = _StackValue.withOffset(stringOffset, ValueType.String, bitWidth);
+    _stack?.add(stackValue);
+    _stringCache?[value] = stackValue;
   }
 
   /// This methods adds a key to a map and should be followed by an add... value call.
@@ -136,8 +135,8 @@ class Builder {
   /// It also implies that you call this method only after you called [startMap].
   void addKey(String value) {
     _integrityCheckOnKeyAddition();
-    if (_keyCache.containsKey(value)) {
-      _stack.add(_keyCache[value]);
+    if (_keyCache?.containsKey(value) == true) {
+      _stack?.add(_keyCache![value]!);
       return;
     }
     final utf8String = utf8.encode(value);
@@ -146,10 +145,9 @@ class Builder {
     final newOffset = _newOffset(length + 1);
     _pushBuffer(utf8String);
     _offset = newOffset;
-    final stackValue =
-        _StackValue.withOffset(keyOffset, ValueType.Key, BitWidth.width8);
-    _stack.add(stackValue);
-    _keyCache[value] = stackValue;
+    final stackValue = _StackValue.withOffset(keyOffset, ValueType.Key, BitWidth.width8);
+    _stack?.add(stackValue);
+    _keyCache?[value] = stackValue;
   }
 
   /// Adds a byte array.
@@ -165,9 +163,8 @@ class Builder {
     final newOffset = _newOffset(length);
     _pushBuffer(value.asUint8List());
     _offset = newOffset;
-    final stackValue =
-        _StackValue.withOffset(blobOffset, ValueType.Blob, bitWidth);
-    _stack.add(stackValue);
+    final stackValue = _StackValue.withOffset(blobOffset, ValueType.Blob, bitWidth);
+    _stack?.add(stackValue);
   }
 
   /// Stores int value indirectly in the buffer.
@@ -178,8 +175,8 @@ class Builder {
   /// By setting the [cache] parameter to true, you make sure that the builder tracks added int value and performs deduplication.
   void addIntIndirectly(int value, {bool cache = false}) {
     _integrityCheckOnValueAddition();
-    if (_indirectIntCache.containsKey(value)) {
-      _stack.add(_indirectIntCache[value]);
+    if (_indirectIntCache?.containsKey(value) == true) {
+      _stack?.add(_indirectIntCache![value]!);
       return;
     }
     final stackValue = _StackValue.withInt(value);
@@ -187,12 +184,11 @@ class Builder {
     final newOffset = _newOffset(byteWidth);
     final valueOffset = _offset;
     _pushBuffer(stackValue.asU8List(stackValue.width));
-    final stackOffset = _StackValue.withOffset(
-        valueOffset, ValueType.IndirectInt, stackValue.width);
-    _stack.add(stackOffset);
+    final stackOffset = _StackValue.withOffset(valueOffset, ValueType.IndirectInt, stackValue.width);
+    _stack?.add(stackOffset);
     _offset = newOffset;
     if (cache) {
-      _indirectIntCache[value] = stackOffset;
+      _indirectIntCache?[value] = stackOffset;
     }
   }
 
@@ -203,8 +199,8 @@ class Builder {
   /// By setting the [cache] parameter to true, you make sure that the builder tracks already added double value and performs deduplication.
   void addDoubleIndirectly(double value, {bool cache = false}) {
     _integrityCheckOnValueAddition();
-    if (cache && _indirectDoubleCache.containsKey(value)) {
-      _stack.add(_indirectDoubleCache[value]);
+    if (cache && _indirectDoubleCache?.containsKey(value) == true) {
+      _stack?.add(_indirectDoubleCache![value]!);
       return;
     }
     final stackValue = _StackValue.withDouble(value);
@@ -212,12 +208,11 @@ class Builder {
     final newOffset = _newOffset(byteWidth);
     final valueOffset = _offset;
     _pushBuffer(stackValue.asU8List(stackValue.width));
-    final stackOffset = _StackValue.withOffset(
-        valueOffset, ValueType.IndirectFloat, stackValue.width);
-    _stack.add(stackOffset);
+    final stackOffset = _StackValue.withOffset(valueOffset, ValueType.IndirectFloat, stackValue.width);
+    _stack?.add(stackOffset);
     _offset = newOffset;
     if (cache) {
-      _indirectDoubleCache[value] = stackOffset;
+      _indirectDoubleCache?[value] = stackOffset;
     }
   }
 
@@ -270,9 +265,7 @@ class Builder {
     tmp._offset = _offset;
     tmp._stack = List.from(_stack);
     tmp._stackPointers = List.from(_stackPointers);
-    tmp._buffer.buffer
-        .asUint8List()
-        .setAll(0, _buffer.buffer.asUint8List(0, _offset));
+    tmp._buffer.buffer.asUint8List().setAll(0, _buffer.buffer.asUint8List(0, _offset));
     for (var i = 0; i < tmp._stackPointers.length; i++) {
       tmp.end();
     }
@@ -288,8 +281,7 @@ class Builder {
     }
     if (_stackPointers.isNotEmpty && _stackPointers.last.isVector == false) {
       if (_stack.last.type != ValueType.Key) {
-        throw StateError(
-            'Adding value to a map before adding a key is prohibited');
+        throw StateError('Adding value to a map before adding a key is prohibited');
       }
     }
   }
@@ -298,15 +290,14 @@ class Builder {
     if (_finished) {
       throw StateError('Adding values after finish is prohibited');
     }
-    if (_stackPointers.isEmpty || _stackPointers.last.isVector) {
+    if (_stackPointers.isEmpty || _stackPointers.last.isVector == true) {
       throw StateError('Adding key before staring a map is prohibited');
     }
   }
 
   void _finish() {
     if (_stack.length != 1) {
-      throw StateError(
-          'Stack has to be exactly 1, but is ${_stack.length}. You have to end all started vectors and maps, before calling [finish]');
+      throw StateError('Stack has to be exactly 1, but is ${_stack.length}. You have to end all started vectors and maps, before calling [finish]');
     }
     final value = _stack[0];
     final byteWidth = _align(value.elementWidth(_offset, 0));
@@ -316,8 +307,7 @@ class Builder {
     _finished = true;
   }
 
-  _StackValue _createVector(int start, int vecLength, int step,
-      [_StackValue keys]) {
+  _StackValue _createVector(int start, int vecLength, int step, [_StackValue? keys]) {
     var bitWidth = BitWidthUtil.uwidth(vecLength);
     var prefixElements = 1;
     if (keys != null) {
@@ -335,7 +325,7 @@ class Builder {
         bitWidth = elemWidth;
       }
       if (i == start) {
-        vectorType = _stack[i].type;
+        vectorType = _stack[i].type!;
         typed &= ValueTypeUtils.isTypedVectorElement(vectorType);
       } else {
         if (vectorType != _stack[i].type) {
@@ -344,9 +334,7 @@ class Builder {
       }
     }
     final byteWidth = _align(bitWidth);
-    final fix = typed & ValueTypeUtils.isNumber(vectorType) &&
-        vecLength >= 2 &&
-        vecLength <= 4;
+    final fix = typed & ValueTypeUtils.isNumber(vectorType) && vecLength >= 2 && vecLength <= 4;
     if (keys != null) {
       _writeStackValue(keys, byteWidth);
       _writeUInt(1 << keys.width.index, byteWidth);
@@ -367,8 +355,7 @@ class Builder {
       return _StackValue.withOffset(vecOffset, ValueType.Map, bitWidth);
     }
     if (typed) {
-      final vType =
-          ValueTypeUtils.toTypedVector(vectorType, fix ? vecLength : 0);
+      final vType = ValueTypeUtils.toTypedVector(vectorType, fix ? vecLength : 0);
       return _StackValue.withOffset(vecOffset, vType, bitWidth);
     }
     return _StackValue.withOffset(vecOffset, ValueType.Vector, bitWidth);
@@ -381,14 +368,16 @@ class Builder {
     _stack.add(vec);
   }
 
-  void _sortKeysAndEndMap(_StackPointer pointer) {
+  void _sortKeysAndEndMap(_StackPointer? pointer) {
+    if (pointer == null) return;
+
     if (((_stack.length - pointer.stackPosition) & 1) == 1) {
       throw StateError(
           'The stack needs to hold key value pairs (even number of elements). Check if you combined [addKey] with add... method calls properly.');
     }
 
     var sorted = true;
-    for (var i = pointer.stackPosition; i < _stack.length - 2; i += 2) {
+    for (var i = (pointer.stackPosition ?? 0); i < _stack.length - 2; i += 2) {
       if (_shouldFlip(_stack[i], _stack[i + 2])) {
         sorted = false;
         break;
@@ -396,7 +385,7 @@ class Builder {
     }
 
     if (sorted == false) {
-      for (var i = pointer.stackPosition; i < _stack.length; i += 2) {
+      for (var i = (pointer.stackPosition ?? 0); i < _stack.length; i += 2) {
         var flipIndex = i;
         for (var j = i + 2; j < _stack.length; j += 2) {
           if (_shouldFlip(_stack[flipIndex], _stack[j])) {
@@ -430,16 +419,14 @@ class Builder {
       keysStackValue = _createVector(pointer.stackPosition, vecLength, 2);
       _keyVectorCache[keysHash] = keysStackValue;
     }
-    final vec =
-        _createVector(pointer.stackPosition + 1, vecLength, 2, keysStackValue);
+    final vec = _createVector(pointer.stackPosition + 1, vecLength, 2, keysStackValue);
     _stack.removeRange(pointer.stackPosition, _stack.length);
     _stack.add(vec);
   }
 
   bool _shouldFlip(_StackValue v1, _StackValue v2) {
     if (v1.type != ValueType.Key || v2.type != ValueType.Key) {
-      throw StateError(
-          'Stack values are not keys $v1 | $v2. Check if you combined [addKey] with add... method calls properly.');
+      throw StateError('Stack values are not keys $v1 | $v2. Check if you combined [addKey] with add... method calls properly.');
     }
 
     var c1, c2;
@@ -467,8 +454,7 @@ class Builder {
       if (byteWidth == 8 || relativeOffset < (1 << (byteWidth * 8))) {
         _writeUInt(relativeOffset, byteWidth);
       } else {
-        throw StateError(
-            'Unexpected size $byteWidth. This might be a bug. Please create an issue https://github.com/google/flatbuffers/issues/new');
+        throw StateError('Unexpected size $byteWidth. This might be a bug. Please create an issue https://github.com/google/flatbuffers/issues/new');
       }
     } else {
       _pushBuffer(value.asU8List(BitWidthUtil.fromByteWidth(byteWidth)));
@@ -537,39 +523,42 @@ class Builder {
 }
 
 class _StackValue {
-  Object _value;
-  int _offset;
-  ValueType _type;
-  BitWidth _width;
+  Object? _value;
+  late int _offset;
+  late ValueType _type;
+  late BitWidth _width;
+
   _StackValue.withNull() {
     _type = ValueType.Null;
     _width = BitWidth.width8;
   }
-  _StackValue.withInt(int value) {
+
+  _StackValue.withInt(int? value) {
     _type = value != null ? ValueType.Int : ValueType.Null;
     _width = BitWidthUtil.width(value);
     _value = value;
   }
-  _StackValue.withBool(bool value) {
+
+  _StackValue.withBool(bool? value) {
     _type = value != null ? ValueType.Bool : ValueType.Null;
     _width = BitWidth.width8;
     _value = value;
   }
-  _StackValue.withDouble(double value) {
+
+  _StackValue.withDouble(double? value) {
     _type = value != null ? ValueType.Float : ValueType.Null;
     _width = BitWidthUtil.width(value);
     _value = value;
   }
-  _StackValue.withOffset(int value, ValueType type, BitWidth width) {
-    _offset = value;
+
+  _StackValue.withOffset(int? value, ValueType type, BitWidth width) {
+    _offset = value??0;
     _type = type;
     _width = width;
   }
 
   BitWidth storedWidth({BitWidth width = BitWidth.width8}) {
-    return ValueTypeUtils.isInline(_type)
-        ? BitWidthUtil.max(_width, width)
-        : _width;
+    return ValueTypeUtils.isInline(_type) ? BitWidthUtil.max(_width, width) : _width;
   }
 
   int storedPackedType({BitWidth width = BitWidth.width8}) {
@@ -580,8 +569,7 @@ class _StackValue {
     if (ValueTypeUtils.isInline(_type)) return _width;
     for (var i = 0; i < 4; i++) {
       final width = 1 << i;
-      final offsetLoc =
-          size + BitWidthUtil.paddingSize(size, width) + index * width;
+      final offsetLoc = size + BitWidthUtil.paddingSize(size, width) + index * width;
       final offset = offsetLoc - _offset;
       final bitWidth = BitWidthUtil.uwidth(offset);
       if (1 << bitWidth.index == width) {
@@ -597,30 +585,30 @@ class _StackValue {
       if (_type == ValueType.Float) {
         if (width == BitWidth.width32) {
           final result = ByteData(4);
-          result.setFloat32(0, _value, Endian.little);
+          result.setFloat32(0, (_value??0.0) as double, Endian.little);
           return result.buffer.asUint8List();
         } else {
           final result = ByteData(8);
-          result.setFloat64(0, _value, Endian.little);
+          result.setFloat64(0, (_value??0.0) as double, Endian.little);
           return result.buffer.asUint8List();
         }
       } else {
         switch (width) {
           case BitWidth.width8:
             final result = ByteData(1);
-            result.setInt8(0, _value);
+            result.setInt8(0, (_value??0) as int);
             return result.buffer.asUint8List();
           case BitWidth.width16:
             final result = ByteData(2);
-            result.setInt16(0, _value, Endian.little);
+            result.setInt16(0, (_value??0) as int, Endian.little);
             return result.buffer.asUint8List();
           case BitWidth.width32:
             final result = ByteData(4);
-            result.setInt32(0, _value, Endian.little);
+            result.setInt32(0, (_value??0) as int, Endian.little);
             return result.buffer.asUint8List();
           case BitWidth.width64:
             final result = ByteData(8);
-            result.setInt64(0, _value, Endian.little);
+            result.setInt64(0, (_value??0) as int, Endian.little);
             return result.buffer.asUint8List();
         }
       }
@@ -632,12 +620,11 @@ class _StackValue {
     }
     if (_type == ValueType.Bool) {
       final result = ByteData(1);
-      result.setInt8(0, _value ? 1 : 0);
+      result.setInt8(0, _value == true ? 1 : 0);
       return result.buffer.asUint8List();
     }
 
-    throw StateError(
-        'Unexpected type: $_type. This might be a bug. Please create an issue https://github.com/google/flatbuffers/issues/new');
+    throw StateError('Unexpected type: $_type. This might be a bug. Please create an issue https://github.com/google/flatbuffers/issues/new');
   }
 
   ValueType get type {
@@ -662,6 +649,7 @@ class _StackValue {
 class _StackPointer {
   int stackPosition;
   bool isVector;
+
   _StackPointer(this.stackPosition, this.isVector);
 }
 
