@@ -8,7 +8,7 @@ import 'package:flutter/services.dart';
 
 class FairJSFairJSDecoderHelper {
   ///提取JS加载地址，assert or http
-  static String transformPath(String beforePath) {
+  static String? transformPath(String? beforePath) {
     if (beforePath == null) {
       // throw FlutterError('');
       return beforePath;
@@ -28,15 +28,15 @@ class FairJSDecoder {
 
   static final FairJSDecoder _fairJSDecoder = FairJSDecoder._internal();
 
-  factory FairJSDecoder({HttpDecoder decoder}) {
+  factory FairJSDecoder({HttpDecoder? decoder}) {
     _httpDecoder = decoder ?? DefaultHttpDecoder();
     return _fairJSDecoder;
   }
 
-  static HttpDecoder _httpDecoder;
+  static HttpDecoder? _httpDecoder;
 
-  Future<String> _resolve(String path) async {
-    if (path.startsWith('http')) {
+  Future<String> _resolve(String? path) async {
+    if (path?.startsWith('http')==true) {
       return await _resolveHttp(path);
     } else {
       //10kb以上资源，会通过isolate的方式去实现，无需额外去操作
@@ -48,14 +48,14 @@ class FairJSDecoder {
   /*
    *解析assert
    */
-  Future<String> _resolveAssert(String assertPath) async =>
-      rootBundle.loadString(assertPath);
+  Future<String> _resolveAssert(String? assertPath) async =>
+      rootBundle.loadString(assertPath??'');
 
   /*
    * 解析Http资源,需要用户自定义自己的解析器
    */
-  Future<String> _resolveHttp(String url) async =>
-      await _httpDecoder.decode(url);
+  Future<String> _resolveHttp(String? url) async =>
+      await _httpDecoder?.decode(url);
 
-  Future<String> decode(String jsPath) => _resolve(jsPath);
+  Future<String> decode(String? jsPath) => _resolve(jsPath);
 }
