@@ -10,17 +10,17 @@ import 'internal/error_tips.dart';
 import 'render/decode.dart';
 
 class Loader {
-  Decoder _decoder;
+  Decoder? _decoder;
 
-  Future<Widget> parse(BuildContext context,
-      {@required String page,
-      String url,
-      @required Map<String, dynamic> data}) async {
+  Future<Widget?> parse(BuildContext context,
+      {required String page,
+      String? url,
+      required Map<String, dynamic> data}) async {
     // Cache decoder for same state instance
     if (_decoder == null) {
       _decoder = Decoder(page, url: url, dataSource: data);
-      await _decoder.resolve(context);
-    } else if(!_decoder.hasResolved){
+      await _decoder!.resolve(context);
+    } else if(!_decoder!.hasResolved){
       // fix bug: in some conditions, _decoder has bean initial, but dataSource
       // hasn't, the waining page will be shown.
       // do nothing, waiting for resolving...
@@ -28,7 +28,7 @@ class Loader {
     }
     Widget widget;
     try {
-      widget = _decoder.toWidget(context);
+      widget = _decoder!.toWidget(context);
       return widget;
     } catch (e) {
       widget = WarningWidget(name: page, url: url, error: e);
