@@ -11,9 +11,9 @@ import 'proxy.dart';
 
 class R {
   final dynamic data;
-  final String exp;
-  static R _empty;
-  final bool needBinding;
+  final String? exp;
+  static R? _empty;
+  final bool? needBinding;
 
   R(this.data, {this.exp, this.needBinding = false});
 
@@ -25,24 +25,24 @@ class R {
 }
 
 abstract class Expression {
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre);
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre);
 
   /// fail-fast
-  bool hitTest(String exp, String pre);
+  bool hitTest(String? exp, String? pre);
 }
 
 class ComponentExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
-    var processed = exp.substring(2, exp.length - 1);
-    var widget = proxy.componentOf(processed);
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
+    var processed = exp?.substring(2, exp.length - 1);
+    var widget = proxy?.componentOf(processed);
     if (widget != null) return R(widget, exp: processed);
-    if (processed.contains('.')) {
-      var s = processed.split('.');
-      assert(s != null && s.length == 2, 'expression is not supported => $exp');
+    if (processed?.contains('.')==true) {
+      var s = processed!.split('.');
+      assert(s.length == 2, 'expression is not supported => $exp');
       var obj = s[0];
       var prop = s[1];
-      var r = proxy.componentOf(obj);
+      var r = proxy!.componentOf(obj);
       if (r is Map) {
         assert(!(r[prop] is Function),
         'should be an instance of widget or const value');
@@ -53,51 +53,51 @@ class ComponentExpression extends Expression {
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp('#\\(.+\\)', multiLine: true).hasMatch(exp);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp('#\\(.+\\)', multiLine: true).hasMatch(exp??'');
   }
 }
 
 class InlineExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
     var regexp = RegExp(r'\$\w+');
-    var matches = regexp.allMatches(pre);
+    var matches = regexp.allMatches(pre??'');
     var builder = _InlineVariableBuilder(
         matches: matches, data: pre, proxyMirror: proxy, binding: binding);
-    binding.addBindValue(builder);
+    binding?.addBindValue(builder);
     return R(builder, exp: exp, needBinding: true);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp(r'\$\w+', multiLine: true).hasMatch(pre);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp(r'\$\w+', multiLine: true).hasMatch(pre??'');
   }
 }
 
 class InlineObjectExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
     var regexp = RegExp(r'\$\{\w.+\}');
-    var matches = regexp.allMatches(pre);
+    var matches = regexp.allMatches(pre??'');
     var builder = _InlineObjectVariableBuilder(
         matches: matches, data: pre, proxyMirror: proxy, binding: binding);
-    binding.addBindValue(builder);
+    binding?.addBindValue(builder);
     return R(builder, exp: exp, needBinding: true);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp(r'\$\{\w.+\}', multiLine: true).hasMatch(pre);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp(r'\$\{\w.+\}', multiLine: true).hasMatch(pre??'');
   }
 }
 
 
 class WidgetParamExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
-    var widgetParameter = exp.substring(9, exp.length - 1);
-    var value = binding.dataOf(widgetParameter);
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
+    var widgetParameter = exp?.substring(9, exp.length - 1);
+    var value = binding?.dataOf(widgetParameter??'');
     if (value != null) {
       return R(value, exp: widgetParameter);
     }
@@ -105,79 +105,79 @@ class WidgetParamExpression extends Expression {
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp('#\\(widget\..+\\)', multiLine: true).hasMatch(exp);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp('#\\(widget\..+\\)', multiLine: true).hasMatch(exp??'');
   }
 }
 
 class ValueExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
     var prop =
-        binding?.bindDataOf(pre) ?? binding.modules.moduleOf(pre)?.call();
+        binding?.bindDataOf(pre??'') ?? binding?.modules?.moduleOf(pre??'')?.call();
     if (prop is ValueNotifier) {
-      var data = _PropBuilder(pre, prop, proxy, binding);
-      binding.addBindValue(data);
+      var data = _PropBuilder(pre??'', prop, proxy, binding);
+      binding?.addBindValue(data);
       return R(data, exp: exp, needBinding: true);
     }
     return R(prop, exp: exp, needBinding: false);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
+  bool hitTest(String? exp, String? pre) {
     return true;
   }
 }
 
 class FunctionExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
     var regexp = RegExp(r'\%\(\w+\)');
-    var matches = regexp.allMatches(exp);
+    var matches = regexp.allMatches(exp??'');
     var builder = _FunctionBuilder(
         matches: matches, data: exp, proxyMirror: proxy, binding: binding);
-    binding.addBindValue(builder);
+    binding?.addBindValue(builder);
     return R(builder, exp: exp, needBinding: false);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp(r'\%\(\w+\)', multiLine: false).hasMatch(exp);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp(r'\%\(\w+\)', multiLine: false).hasMatch(exp??'');
   }
 }
 
 class GestureExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
-    var prop = binding?.bindFunctionOf(exp.substring(2, exp.length - 1));
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
+    var prop = binding?.bindFunctionOf(exp?.substring(2, exp.length - 1)??'');
     return R(prop, exp: exp, needBinding: false);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp(r'\@\(\w+\)', multiLine: false).hasMatch(exp);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp(r'\@\(\w+\)', multiLine: false).hasMatch(exp??'');
   }
 }
 
 class PropValueExpression extends Expression {
   @override
-  R onEvaluate(ProxyMirror proxy, BindingData binding, String exp, String pre) {
-    var expression = exp.substring(2, exp.length - 1);
-    var prop = binding?.bindRuntimeValueOf(expression);
+  R onEvaluate(ProxyMirror? proxy, BindingData? binding, String? exp, String? pre) {
+    var expression = exp?.substring(2, exp.length - 1);
+    var prop = binding?.bindRuntimeValueOf(expression??'');
     return R(prop, exp: expression, needBinding: false);
   }
 
   @override
-  bool hitTest(String exp, String pre) {
-    return RegExp(r'\^\(\w+\)', multiLine: false).hasMatch(exp);
+  bool hitTest(String? exp, String? pre) {
+    return RegExp(r'\^\(\w+\)', multiLine: false).hasMatch(exp??'');
   }
 }
 
-class _BindValueBuilder<T> extends ValueNotifier<T> implements LifeCircle {
-  final String data;
-  final ProxyMirror proxyMirror;
-  final BindingData binding;
-  VoidCallback _listener;
+class _BindValueBuilder<T> extends ValueNotifier<T?> implements LifeCircle {
+  final String? data;
+  final ProxyMirror? proxyMirror;
+  final BindingData? binding;
+  VoidCallback? _listener;
   final List<ValueNotifier> _watchedProps = [];
 
   _BindValueBuilder(this.data, this.proxyMirror, this.binding) : super(null);
@@ -197,19 +197,19 @@ class _BindValueBuilder<T> extends ValueNotifier<T> implements LifeCircle {
   void detach() {
     if (_listener == null) return;
     _watchedProps.forEach((e) {
-      e.removeListener(_listener);
+      e.removeListener(_listener!);
     });
   }
 }
 
 class _InlineVariableBuilder extends _BindValueBuilder<String> {
-  final Iterable<RegExpMatch> matches;
+  final Iterable<RegExpMatch>? matches;
 
   _InlineVariableBuilder(
-      {this.matches, String data, ProxyMirror proxyMirror, BindingData binding})
+      {this.matches, String? data, ProxyMirror? proxyMirror, BindingData? binding})
       : super(data, proxyMirror, binding) {
-    matches.forEach((e) {
-      final bindProp = binding?.bindRuntimeValueOf(e.group(0).substring(1));
+    matches?.forEach((e) {
+      final bindProp = binding?.bindRuntimeValueOf(e.group(0)!.substring(1));
       if (bindProp is ValueNotifier) {
         _watchedProps.add(bindProp);
       }
@@ -220,30 +220,29 @@ class _InlineVariableBuilder extends _BindValueBuilder<String> {
   @override
   String get value {
     var extract = data;
-    matches
-        .map((e) => {
-      '0': binding?.bindRuntimeValueOf(e.group(0).substring(1)),
+    matches?.map((e) => {
+      '0': binding?.bindRuntimeValueOf(e.group(0)?.substring(1)??''),
       '1': e.group(0)
     })
         .forEach((e) {
       var first = e['0'] is ValueNotifier ? e['0'].value : e['0'];
       if (first != null) {
-        extract = extract.replaceFirst(e['1'], '$first');
+        extract = extract?.replaceFirst(e['1'], '$first');
       }
     });
-    return extract;
+    return extract??'';
   }
 }
 
 class _InlineObjectVariableBuilder extends _BindValueBuilder<String> {
-  final Iterable<RegExpMatch> matches;
+  final Iterable<RegExpMatch>? matches;
 
   _InlineObjectVariableBuilder(
-      {this.matches, String data, ProxyMirror proxyMirror, BindingData binding})
+      {this.matches, String? data, ProxyMirror? proxyMirror, BindingData? binding})
       : super(data, proxyMirror, binding) {
-    matches.forEach((e) {
+    matches?.forEach((e) {
       final bindProp =
-      binding?.bindRuntimeValueOf(e.group(0).substring(2, e.group(0).length - 1));
+      binding?.bindRuntimeValueOf(e.group(0)!.substring(2, e.group(0)!.length - 1));
       if (bindProp is ValueNotifier) {
         _watchedProps.add(bindProp);
       }
@@ -255,24 +254,24 @@ class _InlineObjectVariableBuilder extends _BindValueBuilder<String> {
   String get value {
     var extract = data;
     matches
-        .map((e) => {
-      '0': binding?.bindRuntimeValueOf(e.group(0).substring(2, e.group(0).length - 1)),
+        ?.map((e) => {
+      '0': binding?.bindRuntimeValueOf(e.group(0)!.substring(2, e.group(0)!.length - 1)),
       '1': e.group(0)
     })
         .forEach((e) {
       var first = e['0'] is ValueNotifier ? e['0'].value : e['0'];
       if (first != null) {
-        extract = extract.replaceFirst(e['1'], '$first');
+        extract = extract?.replaceFirst(e['1'], '$first');
       }
     });
-    return extract;
+    return extract??'';
   }
 }
 
 
 class _PropBuilder extends _BindValueBuilder {
-  _PropBuilder(String data, ValueNotifier prop, ProxyMirror proxyMirror,
-      BindingData binding)
+  _PropBuilder(String data, ValueNotifier prop, ProxyMirror? proxyMirror,
+      BindingData? binding)
       : super(data, proxyMirror, binding) {
     _watchedProps.add(prop);
     attach();
@@ -280,7 +279,7 @@ class _PropBuilder extends _BindValueBuilder {
 
   @override
   dynamic get value {
-    final prop = binding?.bindDataOf(data);
+    final prop = binding?.bindDataOf(data??'');
     return prop is ValueNotifier ? prop.value : prop;
   }
 }
@@ -302,14 +301,14 @@ class _PropValueBuilder extends _BindValueBuilder {
 }
 
 class _FunctionBuilder extends _BindValueBuilder {
-  final Iterable<RegExpMatch> matches;
+  final Iterable<RegExpMatch>? matches;
 
   _FunctionBuilder(
-      {this.matches, String data, ProxyMirror proxyMirror, BindingData binding})
+      {this.matches, String? data, ProxyMirror? proxyMirror, BindingData? binding})
       : super(data, proxyMirror, binding) {
-    matches.forEach((e) {
+    matches?.forEach((e) {
       var bindProp ;
-      bindProp = binding?.runFunctionOf(e.group(0).substring(2, e.group(0).length - 1));
+      bindProp = binding?.runFunctionOf(e.group(0)!.substring(2, e.group(0)!.length - 1));
       if (bindProp is ValueNotifier) {
         _watchedProps.add(bindProp);
       }
@@ -321,8 +320,8 @@ class _FunctionBuilder extends _BindValueBuilder {
   dynamic get value {
     var extract;
     matches
-        .map((e) => {
-      '0': binding?.runFunctionOf(e.group(0).substring(2, e.group(0).length - 1)),
+        ?.map((e) => {
+      '0': binding?.runFunctionOf(e.group(0)!.substring(2, e.group(0)!.length - 1)),
       '1': e.group(0)
     })
         .forEach((e) {
