@@ -10,14 +10,14 @@ import 'package:flutter/widgets.dart';
 import 'fair_runtime_impl.dart';
 
 abstract class RuntimeFairDelegate {
-  FairState _state;
-  Runtime runtime;
+  FairState? _state;
+  Runtime? runtime;
 
   final _bindValuesMap = <String, PropertyValue>{};
   final _bindFunctionsMap = <String, Function>{};
   final _valueMap = <String, ValueNotifier>{};
 
-  void setRunTime(Runtime runtime) {
+  void setRunTime(Runtime? runtime) {
     this.runtime = runtime;
   }
 
@@ -25,8 +25,8 @@ abstract class RuntimeFairDelegate {
 
   void setState(VoidCallback fn);
 
-  BuildContext get context {
-    return _state.context;
+  BuildContext? get context {
+    return _state?.context;
   }
 
   Map<String, PropertyValue> bindValue() {
@@ -45,7 +45,7 @@ abstract class RuntimeFairDelegate {
   void _bindAllFunc(Map data) {
     data.forEach((key, value) {
       _bindFunctionsMap
-          .addAll({key: () => runtime.invokeMethod(pageName, key, null)});
+          .addAll({key: () => runtime?.invokeMethod(pageName, key, null)});
     });
   }
 
@@ -53,12 +53,12 @@ abstract class RuntimeFairDelegate {
   ///这个地方需要同步调用，因为涉及到数据的绑定，只有将所有数据绑定成功之后，
   ///才能做接下来的操作
   Future<dynamic> bindAll(Map allVariables) async {
-    var map = await runtime.getBindVariableAndFunc(key());
+    var map = await runtime?.getBindVariableAndFunc(key());
 
-    var s = map['result'];
+    var s = map?['result'];
     var d;
-    Map<dynamic, dynamic> variables;
-    List<dynamic> func;
+    Map<dynamic, dynamic>? variables;
+    List<dynamic>? func;
     if (s != null && (d = s['result']) != null) {
       variables = d['variable'];
     }
@@ -67,7 +67,7 @@ abstract class RuntimeFairDelegate {
   }
 
   ///将绑定JS端的变量绑定到当前位置
-  void _bindAllValue(Map data) {
+  void _bindAllValue(Map<dynamic, dynamic>? data) {
     data?.forEach((key, value) {
       _bindValuesMap[key] = () => createValueNotifier(key, value);
     });
@@ -87,9 +87,9 @@ abstract class RuntimeFairDelegate {
 
   void bindBaseFunc() {
     _bindFunctionsMap['runtimeInvokeMethod'] = (funcName, [props]) {
-      var arguments = null;
+      var arguments;
       if (props != null) {
-        arguments = List();
+        arguments = [];
         arguments.add(props);
       }
       return runtime?.invokeMethod(pageName, funcName, arguments);
