@@ -33,16 +33,16 @@ import 'type.dart';
 class FairWidget extends StatefulWidget {
   /// Unique widget name. If this name is registered with [FairApp.widget], the [path] can be omitted.
   /// If the same bundle is reuse with different data the name should appended with suffix, eg: $name-$data.hashcode
-  final String name;
+  final String? name;
 
   /// Bundle path asset|url.
   /// Also see:
   ///
   /// * [name], unique name binds to this FairWidget
-  final String path;
+  final String? path;
 
   /// Optional, data source relate to this FairWidget.
-  final Map<String, dynamic> data;
+  final Map<String, dynamic>? data;
 
   /// Optional, provide the loading widget before real content is ready.
   /// Will use the placeholder configured with [FairApp] if it's not provided.
@@ -50,32 +50,29 @@ class FairWidget extends StatefulWidget {
   /// Also see:
   ///
   /// * [FairApp.placeholderBuilder], global placeholder for [FairApp]
-  final WidgetBuilder holder;
+  final WidgetBuilder? holder;
 
   /// Optional delegate for the fair bundle
   /// Also see:
   ///
   /// * delegate of [FairApp]
-  final FairDelegate delegate;
+  final FairDelegate? delegate;
 
   /// Work with [AutomaticKeepAliveClientMixin], usually for TabBar
-  final bool wantKeepAlive;
+  final bool? wantKeepAlive;
 
   FairWidget({
-    Key key,
+    Key? key,
     this.name,
     this.path,
     this.data,
     this.holder,
     this.delegate,
     this.wantKeepAlive = false,
-  })  : assert(!(name == null && path == null),
-            'FairWidget require a global registered `name` or bundle `path`'),
+  })  : assert(!(name == null && path == null), 'FairWidget require a global registered `name` or bundle `path`'),
         assert(() {
           if (data == null) return true;
-          if (!(data is Map &&
-              data.values
-                  .every((e) => e is int || e is double || e is String))) {
+          if (!(data is Map && data.values.every((e) => e is int || e is double || e is String))) {
             log('data must be a map of primary value such as int, double or String. Object reference can be broken!');
           }
           return true;
@@ -96,23 +93,20 @@ class FairWidget extends StatefulWidget {
   }
 }
 
-class FairState extends State<FairWidget>
-    with Loader, AutomaticKeepAliveClientMixin<FairWidget>
-    implements FairMessageCallback<String> {
-  Widget _child;
-  FairApp _fairApp;
-  String bundleType;
-  String state2key;
+class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientMixin<FairWidget> implements FairMessageCallback<String> {
+  Widget? _child;
+  FairApp? _fairApp;
+  String? bundleType;
+  late String state2key;
 
   // None nullable
-  FairDelegate delegate;
+  late FairDelegate delegate;
 
   @override
   void initState() {
     super.initState();
     state2key = GlobalState.id(widget.name);
-    delegate = widget.delegate ??
-        GlobalState.of(widget.name).call(context, widget.data);
+    delegate = widget.delegate ?? GlobalState.of(widget.name).call(context, widget.data);
     delegate._bindState(this);
     delegate.initState();
   }
@@ -122,34 +116,32 @@ class FairState extends State<FairWidget>
     super.didChangeDependencies();
     _fairApp ??= FairApp.of(context);
     //加载js的文件地址
-    _resolveFairRes(
-        _fairApp, FairJSFairJSDecoderHelper.transformPath(widget.path));
+    _resolveFairRes(_fairApp!, FairJSFairJSDecoderHelper.transformPath(widget.path));
   }
 
-  Future<dynamic> _resolveFairRes(FairApp _mFairApp, String jsPath) async {
-    var resolveJS =
-        await FairJSDecoder(decoder: _mFairApp.httpDecoder).decode(jsPath);
-    await Future.wait([
-      _mFairApp.runtime.addScript(state2key, resolveJS, widget.data),
-      _mFairApp.register(this)
-    ]);
-    delegate?.didChangeDependencies();
+  Future<dynamic> _resolveFairRes(FairApp _mFairApp, String? jsPath) async {
+    var resolveJS='';
+    try {
+      resolveJS = await FairJSDecoder(decoder: _mFairApp.httpDecoder).decode(jsPath);
+    } catch (e) {
+      print(e);
+    }
+    await Future.wait([_mFairApp.runtime.addScript(state2key, resolveJS, widget.data), _mFairApp.register(this)]);
+    delegate.didChangeDependencies();
     _reload();
   }
 
   @override
   void didUpdateWidget(covariant FairWidget oldWidget) {
     super.didUpdateWidget(oldWidget);
-    delegate?.didUpdateWidget(oldWidget);
+    delegate.didUpdateWidget(oldWidget);
   }
 
   void _reload() {
     var name = state2key;
-    var path = widget.path ?? _fairApp.pathOfBundle(widget.name);
-    bundleType = widget.path != null && widget.path.startsWith('http')
-        ? 'Http'
-        : 'Asset';
-    parse(context, page: name, url: path, data: widget.data).then((value) {
+    var path = widget.path ?? _fairApp!.pathOfBundle(widget.name??'');
+    bundleType = widget.path != null && widget.path?.startsWith('http')==true ? 'Http' : 'Asset';
+    parse(context, page: name, url: path, data: widget.data??{}).then((value) {
       if (mounted && value != null) {
         setState(() => _child = value);
       }
@@ -160,22 +152,22 @@ class FairState extends State<FairWidget>
   Widget build(BuildContext context) {
     super.build(context);
     assert(_fairApp != null, 'FairWidget must be descendant of FairApp');
-    var builder = widget.holder ?? _fairApp.placeholderBuilder;
-    var result = _child ?? builder(context);
-    if (!kReleaseMode && _fairApp.debugShowFairBanner) {
-      result = _CheckedModeBanner(bundleType, child: result);
+    var builder = widget.holder ?? _fairApp?.placeholderBuilder;
+    var result = _child ?? builder?.call(context);
+    if (!kReleaseMode && _fairApp!.debugShowFairBanner) {
+      result = _CheckedModeBanner(bundleType??'', child: result??Container());
     }
-    return result;
+    return result??Container();
   }
 
   @override
-  bool get wantKeepAlive => widget.wantKeepAlive;
+  bool get wantKeepAlive => widget.wantKeepAlive??false;
 
   @override
   void dispose() {
     super.dispose();
     _fairApp?.unregister(this);
-    delegate?.dispose();
+    delegate.dispose();
   }
 
   // String get state2key =>
@@ -183,7 +175,13 @@ class FairState extends State<FairWidget>
 
   @override
   void call(String t) {
-    delegate.notifyValue(jsonDecode(t));
+    var params={};
+    try {
+      params= jsonDecode(t);
+    } catch (e) {
+      print(e);
+    }
+    delegate.notifyValue(params);
   }
 
   @override
@@ -192,12 +190,12 @@ class FairState extends State<FairWidget>
 
 /// Delegate for business logic. The delegate share similar life-circle with [State].
 class FairDelegate extends RuntimeFairDelegate {
-  FairState _state;
-  String _key;
+  FairState? _state;
+ late String _key;
 
-  void _bindState(FairState state) {
+  void _bindState(FairState? state) {
     assert(state != null, 'FairState should not be null');
-    _state = state;
+    _state = state!;
     _key = state.state2key;
   }
 
@@ -205,14 +203,15 @@ class FairDelegate extends RuntimeFairDelegate {
   /// Usually this can cost several milliseconds depend on the complexity of DSL.
   @override
   void setState(VoidCallback fn) {
-    if (_state == null || !_state.mounted) return;
+    if (_state == null || !_state!.mounted) return;
     // ignore: invalid_use_of_protected_member
-    _state.setState(fn);
-    _state._reload();
+    _state?.setState(fn);
+    _state?._reload();
   }
 
+  @override
   BuildContext get context {
-    return _state.context;
+    return _state!.context;
   }
 
   @override
@@ -257,8 +256,7 @@ class FairDelegate extends RuntimeFairDelegate {
 /// Does nothing in release mode.
 class _CheckedModeBanner extends StatelessWidget {
   /// Creates a const checked banner.
-  const _CheckedModeBanner(this.appendix, {Key key, @required this.child})
-      : super(key: key);
+  const _CheckedModeBanner(this.appendix, {Key? key, required this.child}) : super(key: key);
 
   /// The widget to show behind the banner.
   ///
@@ -271,11 +269,11 @@ class _CheckedModeBanner extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var result = Banner(
-      child: child,
       message: 'Fair $appendix',
       textDirection: TextDirection.ltr,
       location: BannerLocation.topStart,
       color: appendix == 'Http' ? Colors.green : Colors.orange,
+      child: child,
     );
     return result;
   }
