@@ -7,14 +7,14 @@
 import '../public_type.dart';
 
 class FairModuleRegistry {
-  final _modules = <String, FairModuleBuilder>{};
+  final Map _modules = <String, FairModuleBuilder>{};
 
-  void addAll(Map<String, FairModuleBuilder> modules) {
+  void addAll(Map<String, FairModuleBuilder>? modules) {
     if (modules == null || modules.isEmpty) {
       return;
     }
     _modules.addAll(modules);
   }
 
-  FairModuleBuilder moduleOf(String name) => _modules[name];
+  FairModuleBuilder? moduleOf(String name) => _modules[name];
 }
